==> code/1.Patient_Covariates.ipynb <==
library(bigrquery)
library(tidyverse)
library(labelled)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# pull covariates from tbl_SRCode
# cholesterol
chol <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text, src_numericvalue, src_numericunit FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode`
WHERE src_ctv3code LIKE '%XaIRD%' OR src_ctv3code LIKE '%XaJe9%' OR src_ctv3code LIKE '%Xa86v%' OR src_ctv3code LIKE '%XSK14%' OR src_ctv3code LIKE '%44PF.%' OR src_ctv3code LIKE '%Y1cc2%'"
tb1 <- bq_project_query(projectid, chol)
chol <- bq_table_download(tb1)

# systolic blood pressure
sbp <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text, src_numericvalue, src_numericunit FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode`
WHERE src_ctv3code LIKE '%XaIwj%' OR src_ctv3code LIKE '%XaF4F%' OR src_ctv3code LIKE '%XaJ2E%'"
tb2 <- bq_project_query(projectid, sbp)
sbp <- bq_table_download(tb2)

# diastolic blood pressure
dbp <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text, src_numericvalue, src_numericunit FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode`
WHERE src_ctv3code LIKE '%XaF4S%' OR src_ctv3code LIKE '%XaIwk%' OR src_ctv3code LIKE '%XaJ2F%'"
tb21 <- bq_project_query(projectid, dbp)
dbp <- bq_table_download(tb21)

# BMI
bmi <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text, src_numericvalue, src_numericunit FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` WHERE src_ctv3code LIKE '%22K..%'"
tb3 <- bq_project_query(projectid, bmi)
bmi <- bq_table_download(tb3)

# Smoking
smoke <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode`
WHERE src_ctv3code LIKE '%Ub1na%' OR src_ctv3code LIKE '%Xe0oq%' OR src_ctv3code LIKE '%137R.%' OR src_ctv3code LIKE '%Ub0oq%'"
tb4 <- bq_project_query(projectid, smoke)
smoke <- bq_table_download(tb4)

# Renal disease
# Creatinine
renal_cr <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text, src_numericvalue, src_numericunit FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode`
WHERE src_ctv3code LIKE '%44J3.%' OR src_ctv3code LIKE '%46M7.%' OR src_ctv3code LIKE '%X771Q%' OR src_ctv3code LIKE '%X80D7%' 
OR src_ctv3code LIKE '%XE26a%' OR src_ctv3code LIKE '%XE2q5%' OR src_ctv3code LIKE '%XE2qO%' OR src_ctv3code LIKE '%XaETQ%'"
tb5 <- bq_project_query(projectid, renal_cr)
renal_cr <- bq_table_download(tb5)

# Albumin
renal_al <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text, src_numericvalue, src_numericunit FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode`
WHERE src_ctv3code LIKE '%46TC.%' OR src_ctv3code LIKE '%X773Y%' OR src_ctv3code LIKE '%XE2n3%' OR src_ctv3code LIKE '%XE2n4%'"
tb51 <- bq_project_query(projectid, renal_al)
renal_al <- bq_table_download(tb51)

# eGFR
renal_egfr <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text, src_numericvalue, src_numericunit FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode`
WHERE src_ctv3code LIKE '%XaZpN%'"
tb52 <- bq_project_query(projectid, renal_egfr)
renal_egfr <- bq_table_download(tb52)

nrow(chol)
nrow(bmi)
nrow(smoke)
nrow(sbp)
nrow(dbp)
nrow(renal_cr)
nrow(renal_al)
nrow(renal_egfr)

head(chol)
table(chol$src_numericunit)
table(chol$src_ctv3text, chol$src_numericunit)

cholesterol <- chol %>% filter(src_numericunit == "mmol/L" | src_numericunit == "mmol/l") # keep only mmol/L units
sbp$src_numericvalue <- as.numeric(sbp$src_numericvalue)
sysbp <- sbp %>% filter(src_numericunit == "mmHg") %>% mutate(src_numericvalue = round(src_numericvalue, digits = 0)) # keep only mmHg units
dbp$src_numericvalue <- as.numeric(dbp$src_numericvalue)
diabp <- dbp %>% filter(src_numericunit == "mmHg") %>% mutate(src_numericvalue = round(src_numericvalue, digits = 0)) # keep only mmHg units
BMI <- bmi %>% filter(src_numericunit == "Kg/mï¿½" | src_numericunit == "Kg/m��«") %>% mutate(src_numericunit = "kg/m2")
renal_CR <- renal_cr %>% filter(src_numericunit == "umol/L")
renal_AL <- renal_al %>% filter(src_numericunit == "mg/mmol")
renal_EGFR <- renal_egfr %>% filter(src_numericunit == "mL/min/1.73m*2")

nrow(cholesterol)
nrow(BMI)
nrow(sysbp)
nrow(diabp)
nrow(renal_CR)
nrow(renal_AL)
nrow(renal_EGFR)

colnames(cholesterol)
table(cholesterol$src_ctv3text, cholesterol$src_numericunit)
cholesterol %>% select(src_numericunit, src_numericvalue, src_ctv3text)

# Clean covariate tables
# Cholesterol
chol$src_numericvalue <- as.numeric(chol$src_numericvalue)
cholesterol <- cholesterol %>%
    filter(!src_ctv3text == "Cervical smear overdue") %>%
  mutate(cholesterol = case_when(
    src_numericvalue <= 5 ~ 1,
    src_numericvalue > 5 ~ 2)) %>%
  set_value_labels(cholesterol = c("Healthy" = 1, "High" = 2)) %>%
  select(person_id, date_cholesterol = src_dateevent, cholesterol)
cholesterol$date_cholesterol <- as.Date(cholesterol$date_cholesterol)
cholesterol <- cholesterol[!duplicated(cholesterol),]
nrow(cholesterol)
table(cholesterol$cholesterol)

# Clean covariate tables
# BMI
BMI$src_dateevent <- as.Date(BMI$src_dateevent)
BMI$src_numericvalue <- as.numeric(BMI$src_numericvalue)
# two decimal places
BMI <- BMI %>%
  mutate(bmi = case_when(
    src_numericvalue < 18.5 ~ 1,
    src_numericvalue >= 18.5 & src_numericvalue <= 24.9 ~ 2,
    src_numericvalue >= 25 & src_numericvalue <= 29.9 ~ 3,
    src_numericvalue >= 30 ~ 4)) %>%
  set_value_labels(bmi = c("Underweight" = 1, "Healthy weight" = 2, "Overweight" = 3, "Obese" = 4)) %>%
  select(person_id, date_bmi = src_dateevent, bmi, bmi_value = src_numericvalue)
BMI <- BMI[!duplicated(BMI),]
nrow(BMI)
table(BMI$bmi)


sum(duplicated(BMI))

# Clean covariate tables
# Smoking
smoke$src_dateevent <- as.Date(smoke$src_dateevent)
smoking <- smoke %>%
  mutate(smoking = case_when(
    src_ctv3text == "Smoker" ~ 1,
    src_ctv3text == "Ex-smoker" ~ 2,
    src_ctv3text == "Non-smoker" ~ 3)) %>%
  set_value_labels(smoking = c("Smoker" = 1, "Ex-smoker" = 2, "Non-smoker" = 3)) %>%
  select(person_id, date_smoking = src_dateevent, smoking)
smoking <- smoking[!duplicated(smoking),]
nrow(smoking)
table(smoking$smoking)

# Clean covariate tables
# Systolic Blood pressure
sysbp$src_dateevent <- as.Date(sysbp$src_dateevent)
sysbp <- sysbp %>%
  mutate(sys_bp = case_when(
    src_numericvalue < 90 ~ 1,
    src_numericvalue >= 90 & src_numericvalue <= 139 ~ 2,
    src_numericvalue >= 140 ~ 3)) %>%
  set_value_labels(sys_bp = c("Low" = 1, "Normal" = 2, "High" = 3)) %>%
  select(person_id, date_bp = src_dateevent, sys_bp, sbp_value = src_numericvalue)
table(sysbp$sys_bp)

sysbp <- sysbp[!duplicated(sysbp),]
nrow(sysbp)

table(sysbp$sys_bp)

# Diastolic Blood pressure
diabp$src_dateevent <- as.Date(diabp$src_dateevent)
diabp <- diabp %>%
  mutate(dia_bp = case_when(
    src_numericvalue < 60 ~ 1,
    src_numericvalue >= 60 & src_numericvalue <= 89 ~ 2,
    src_numericvalue >= 90 ~ 3)) %>%
  set_value_labels(dia_bp = c("Low" = 1, "Normal" = 2, "High" = 3)) %>%
  select(person_id, date_bp = src_dateevent, dia_bp, dbp_value = src_numericvalue)
table(diabp$dia_bp)

diabp <- diabp[!duplicated(diabp),]
nrow(diabp)

table(diabp$dia_bp)

# Join systolic and diastolic blood pressure tables together
bp <- full_join(sysbp, diabp, by = c("person_id", "date_bp"))
head(bp)
bp <- bp[!duplicated(bp),]
nrow(bp)
table(bp$sys_bp, bp$dia_bp)

# BP variable
bp <- bp %>%
  mutate(bp = case_when(
    sys_bp == 1 & dia_bp == 1 ~ 1,
    sys_bp == 1 | dia_bp == 1 ~ 11,
    sys_bp == 2 & dia_bp == 2 ~ 2,
    sys_bp == 3 & dia_bp == 3 ~ 3,
    sys_bp == 3 | dia_bp == 3 ~ 31,
    sys_bp == 1 & dia_bp == 3 ~ 32,
    sys_bp == 3 & dia_bp == 1 ~ 33,
    is.na(sys_bp) | is.na(dia_bp) ~ NA)) %>%
  mutate(blood_pressure = case_when(
    bp == 1 | bp == 11 ~ 1,
    bp == 2 ~ 2,
    bp == 3 | bp == 31 | bp == 32 | bp == 33 ~ 3,
    is.na(bp) ~ NA)) %>%
  set_value_labels(blood_pressure = c("Low" = 1, "Normal" = 2, "High" = 3)) %>%
  select(person_id, date_bp, blood_pressure, sbp_value, dbp_value)
table(bp$blood_pressure)
bp <- bp[!duplicated(bp),]
nrow(bp)
table(bp$blood_pressure)

# Clean covariate tables
# Kidney Function
#renal_CR$src_dateevent <- as.Date(renal_CR$src_dateevent)
#renal_AL$src_dateevent <- as.Date(renal_AL$src_dateevent)
#renal_EGFR$src_dateevent <- as.Date(renal_EGFR$src_dateevent)

# Creatinine
renal_CR <- renal_CR %>%
    filter(!src_numericvalue < 0) %>%
    mutate(creatinine = case_when(
        src_numericvalue <= 100 ~ 1,
        src_numericvalue > 100 & src_numericvalue < 500 ~ 2,
        src_numericvalue >= 500 ~ 3)) %>%
    set_value_labels(creatinine = c("Normal" = 1, "High" = 2, "Very High" = 3)) %>%
    select(person_id, date_creatinine = src_dateevent, creatinine)
renal_CR <- renal_CR[!duplicated(renal_CR),]
nrow(renal_CR)
table(renal_CR$creatinine)

# Albumin:Creatinine ratio (ACR)
renal_AL <- renal_AL %>%
    mutate(acr = case_when(
        src_numericvalue <= 30 ~ 1,
        src_numericvalue > 30 & src_numericvalue < 300 ~ 2,
        src_numericvalue >= 300 ~ 3)) %>%
    set_value_labels(acr = c("Normal" = 1, "High" = 2, "Very High" = 3)) %>%
    select(person_id, date_acr = src_dateevent, acr)
renal_AL <- renal_AL[!duplicated(renal_AL),]
nrow(renal_AL)
table(renal_AL$acr)

# eGFR
renal_EGFR <- renal_EGFR %>%
    mutate(egfr = case_when(
        src_numericvalue >= 90 ~ 1,
        src_numericvalue < 90 ~ 2)) %>%
    set_value_labels(egfr = c("Normal" = 1, "Reduced kidney function" = 2)) %>%
    select(person_id, date_egfr = src_dateevent, egfr)
renal_EGFR <- renal_EGFR[!duplicated(renal_EGFR),]
nrow(renal_EGFR)
table(renal_EGFR$egfr)

colnames(bp)
cholesterol$cholesterol <- as.integer(cholesterol$cholesterol)
BMI$bmi <- as.integer(BMI$bmi)
smoking$smoking <- as.integer(smoking$smoking)
bp$blood_pressure <- as.integer(bp$blood_pressure)
#renal$renal_disease <- as.integer(renal$renal_disease)
renal_CR$creatinine <- as.integer(renal_CR$creatinine)
renal_AL$acr <- as.integer(renal_AL$acr)
renal_EGFR$egfr <- as.integer(renal_EGFR$egfr)

# Make sure date variables are class 'Date'
cholesterol$date_cholesterol <- as.Date(cholesterol$date_cholesterol)
BMI$date_bmi <- as.Date(BMI$date_bmi)
smoking$date_smoking <- as.Date(smoking$date_smoking)
bp$date_bp <- as.Date(bp$date_bp)
renal_CR$date_creatinine <- as.Date(renal_CR$date_creatinine)
renal_AL$date_acr <- as.Date(renal_AL$date_acr)
renal_EGFR$date_egfr <- as.Date(renal_EGFR$date_egfr)

head(cholesterol)

# keep dates only between 2006-2018 i.e. a maximum of 1 year before/after diabetes diagnosis
cholesterol <- cholesterol %>%
    mutate(within_date = ifelse(date_cholesterol >= "2006-01-01" & date_cholesterol <= "2018-12-31",1,0))
BMI <- BMI %>%
    mutate(within_date = ifelse(date_bmi >= "2006-01-01" & date_bmi <= "2018-12-31",1,0))
smoking <- smoking %>%
    mutate(within_date = ifelse(date_smoking >= "2006-01-01" & date_smoking <= "2018-12-31",1,0))
bp <- bp %>%
    mutate(within_date = ifelse(date_bp >= "2006-01-01" & date_bp <= "2018-12-31",1,0))
renal_CR <- renal_CR %>%
    mutate(within_date = ifelse(date_creatinine >= "2006-01-01" & date_creatinine <= "2018-12-31",1,0))
renal_AL <- renal_AL %>%
    mutate(within_date = ifelse(date_acr >= "2006-01-01" & date_acr <= "2018-12-31",1,0))
renal_EGFR <- renal_EGFR %>%
    mutate(within_date = ifelse(date_egfr >= "2006-01-01" & date_egfr <= "2018-12-31",1,0))

table(cholesterol$within_date)
table(BMI$within_date)
table(smoking$within_date)
table(bp$within_date)
table(renal_CR$within_date)
table(renal_AL$within_date)
table(renal_EGFR$within_date)

# Keep only those within the dates
cholesterol <- cholesterol %>%
    filter(!within_date == 0) %>%
    select(-within_date)
BMI <- BMI %>%
    filter(!within_date == 0) %>%
    select(-within_date)
smoking <- smoking %>%
    filter(!within_date == 0) %>%
    select(-within_date)
bp <- bp %>%
    filter(!within_date == 0) %>%
    select(-within_date)
renal_CR <- renal_CR %>%
    filter(!within_date == 0) %>%
    select(-within_date)
renal_AL <- renal_AL %>%
    filter(!within_date == 0) %>%
    select(-within_date)
renal_EGFR <- renal_EGFR %>%
    filter(!within_date == 0) %>%
    select(-within_date)

head(cholesterol)

nrow(cholesterol)
nrow(BMI)
nrow(smoking)
nrow(bp)
nrow(renal_CR)
nrow(renal_AL)
nrow(renal_EGFR)

bq_auth()

# save each dataset
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
# Cholesterol
dest_table <- bq_table(dest_dataset, "Cholesterol")
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("date_cholesterol", "DATE"),
                              bq_field("cholesterol", "INT64")))
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
bq_table_upload(dest_table, cholesterol, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')
# BMI
dest_table <- bq_table(dest_dataset, "BMI")
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("date_bmi", "DATE"),
                              bq_field("bmi", "INT64"),
                              bq_field("bmi_value", "FLOAT64")))
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
bq_table_upload(dest_table, BMI, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')
# Smoking
dest_table <- bq_table(dest_dataset, "Smoking")
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("date_smoking", "DATE"),
                              bq_field("smoking", "INT64")))
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
bq_table_upload(dest_table, smoking, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')
# Blood Pressure
dest_table <- bq_table(dest_dataset, "Blood_Pressure")
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("date_bp", "DATE"),
                              bq_field("blood_pressure", "INT64"),
                              bq_field("sbp_value", "FLOAT64"),
                              bq_field("dbp_value", "FLOAT64")))
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
bq_table_upload(dest_table, bp, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')

# Renal disease: Creatinine
dest_table <- bq_table(dest_dataset, "Creatinine")
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("date_creatinine", "DATE"),
                              bq_field("creatinine", "INT64")))
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
bq_table_upload(dest_table, renal_CR, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')

# Renal disease: ACR
dest_table <- bq_table(dest_dataset, "ACR")
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("date_acr", "DATE"),
                              bq_field("acr", "INT64")))
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
bq_table_upload(dest_table, renal_AL, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')

# Renal disease: eGFR
dest_table <- bq_table(dest_dataset, "eGFR")
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("date_egfr", "DATE"),
                              bq_field("egfr", "INT64")))
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
bq_table_upload(dest_table, renal_EGFR, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')

colnames(BMI)
class(BMI$bmi_value)
colnames(bp)
