==> code/1.Patient_Denom.ipynb <==
library(bigrquery)
library(tidyverse)
#install.packages("labelled")
library(labelled)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# Load person table
sql1 <- "SELECT person_id, birth_datetime, death_datetime, gender_source_value, ethnicity_concept_id FROM `yhcr-prd-phm-bia-core.CB_FDM_MASTER.person`
WHERE ethnicity_concept_id = 46285833
OR ethnicity_concept_id = 46285834
OR ethnicity_concept_id = 46285831
OR ethnicity_concept_id = 46285832
OR ethnicity_concept_id = 46285835
OR ethnicity_concept_id = 46286811
OR ethnicity_concept_id = 46285836
OR ethnicity_concept_id = 46285837
OR ethnicity_concept_id = 46285829
OR ethnicity_concept_id = 46285839
OR ethnicity_concept_id = 46285828
OR ethnicity_concept_id = 46285827
OR ethnicity_concept_id = 46285830
OR ethnicity_concept_id = 46285838
OR ethnicity_concept_id = 46285839
OR ethnicity_concept_id = 46286810
OR ethnicity_concept_id = 46285825
OR ethnicity_concept_id = 46285824
OR ethnicity_concept_id = 46285826
OR ethnicity_concept_id = 46286810"

tb1 <- bq_project_query(projectid, sql1)

pat1 <-bq_table_download(tb1) # store data in tibble

nrow(pat1)

patient <- pat1 %>%
    mutate(gender = case_when(
        gender_source_value == "M" | gender_source_value == "Male" ~ 1,
        gender_source_value == "F" | gender_source_value == "Female" ~ 2)) %>%
    select(-gender_source_value) %>%
  mutate(ethnicity = case_when(
    ethnicity_concept_id == "46286810" | ethnicity_concept_id == "46285826" | ethnicity_concept_id == "46285825" | ethnicity_concept_id == "46285824" ~ 1,
    ethnicity_concept_id == "46285832" | ethnicity_concept_id == "46285833" | ethnicity_concept_id == "46285831" ~ 2,
    TRUE ~ 3)) %>%
select(-ethnicity_concept_id)

table(patient$gender)
table(patient$ethnicity)

# Load imd by lsoa, and person lsoa
lsoa <- "SELECT * FROM `yhcr-prd-phm-bia-core.CB_LOOKUPS.tbl_person_lsoa`"
tb <- bq_project_query(projectid, lsoa)
lsoa <-bq_table_download(tb) 

imd_lsoa <- "SELECT LSOA_code, Index_of_Multiple_Deprivation_Decile FROM `yhcr-prd-phm-bia-core.CB_LOOKUPS.tbl_IMD_by_LSOA`"
tb1 <- bq_project_query(projectid, imd_lsoa)
imd_lsoa <- bq_table_download(tb1)

imd_lsoa <- imd_lsoa %>%
rename(lsoa = LSOA_code) %>%
rename(imd_decile = Index_of_Multiple_Deprivation_Decile)
head(imd_lsoa)
head(lsoa)

imd_person <- full_join(lsoa, imd_lsoa, by = "lsoa")
imd_person <- imd_person[!duplicated(imd_person$person_id),]
imd_person <- imd_person %>% select(-lsoa)
head(imd_person)
sum(is.na(imd_person$imd_decile))
imd_person <- imd_person[!is.na(imd_person$imd_decile),]
nrow(imd_person)

table(imd_person$imd_decile)
384207+105922
128678+104455
78357+95040
66898+60152
48298+26930

pat_imd <- left_join(patient, imd_person, by = "person_id")

pat_imd <- pat_imd %>%
    filter(!duplicated(pat_imd))

pat_imd <- pat_imd %>%
    filter(!is.na(imd_decile))

pat_imd <- pat_imd %>%
    filter(!is.na(gender))

head(pat_imd)
nrow(pat_imd)
patient <- pat_imd

patient$birth_datetime <- as.Date(patient$birth_datetime)
patient$death_datetime <- as.Date(patient$death_datetime)

head(patient)
sum(duplicated(patient))
patient <- patient %>% select(-datasetref)

table(patient$gender)
table(patient$ethnicity)
table(patient$imd_decile)
290711+77500
93078+74043
53405+63714
42832+40513
30533+17857

bq_auth()

# set destination table - do this regardless of whether it already exists
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
dest_table <- bq_table(dest_dataset, "Patient_Denom")

# create a full table spec as follows
# then pass this to bq_table_upload
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("birth_datetime", "DATE"),
                              bq_field("death_datetime", "DATE"),
                              bq_field("gender", "INT64"),
                              bq_field("ethnicity", "INT64"),
                              bq_field("imd_decile", "INT64")))

# delete an existing table if necessary
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
# upload file_data to dest_table
bq_table_upload(dest_table, patient, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')
