==> code/3.T2D_PrimaryCare.ipynb <==
library(bigrquery)
library(tidyverse)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# Set your query
sql <- "SELECT * FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` 
WHERE src_dateevent BETWEEN '2007-01-01' AND '2022-12-31' AND (src_ctv3code LIKE '%C1011%' OR
src_ctv3code LIKE '%C1031%' OR
src_ctv3code LIKE '%C1090%' OR
src_ctv3code LIKE '%C1091%' OR
src_ctv3code LIKE '%C1092%' OR
src_ctv3code LIKE '%C1093%' OR
src_ctv3code LIKE '%C1094%' OR
src_ctv3code LIKE '%C1095%' OR
src_ctv3code LIKE '%C1096%' OR
src_ctv3code LIKE '%C1097%' OR
src_ctv3code LIKE '%L1806%' OR
src_ctv3code LIKE '%X40J5%' OR
src_ctv3code LIKE '%X40J6%' OR
src_ctv3code LIKE '%XaELQ%' OR
src_ctv3code LIKE '%XaEnp%' OR
src_ctv3code LIKE '%XaEnq%' OR
src_ctv3code LIKE '%XaF05%' OR
src_ctv3code LIKE '%XaFWI%' OR
src_ctv3code LIKE '%XaFmA%' OR
src_ctv3code LIKE '%XaFn7%' OR
src_ctv3code LIKE '%XaFn8%' OR
src_ctv3code LIKE '%XaFn9%' OR
src_ctv3code LIKE '%XaIfG%' OR
src_ctv3code LIKE '%XaIfI%' OR
src_ctv3code LIKE '%XaIrf%' OR
src_ctv3code LIKE '%XaIzQ%' OR
src_ctv3code LIKE '%XaIzR%' OR
src_ctv3code LIKE '%XaJQp%' OR
src_ctv3code LIKE '%XaKyX%')"

# Run the query; this returns a bq_table object that you can query further
tb <- bq_project_query(projectid, sql)

# Store data in a tibble
t2d <-bq_table_download(tb)
# Print table
head(t2d)

# keep on columns 1,4,7 and 8
t2d <- t2d %>% select(1,4,7,8)
nrow(t2d)
sum(is.na(t2d))
sum(duplicated(t2d))

# remove duplicated rows and count remaining number of rows
t2d <- t2d[!duplicated(t2d),]
nrow(t2d)

# add indicator variable for t2d and remove ctv3 cols
t2d <- t2d %>%
  mutate(t2d_status = 1) %>%
  select(-3,-4)

# change class
t2d$src_dateevent <- as.Date(t2d$src_dateevent)
t2d$t2d_status <- as.integer(t2d$t2d_status)

head(t2d)
sum(duplicated(t2d$person_id))

# choose earliest date of event per person
t2d <- t2d %>%
group_by(person_id) %>%
arrange(src_dateevent) %>%
mutate(event_number = row_number()) %>%
slice_min(event_number) %>%
select(-event_number)
nrow(t2d)

# set destination table - do this regardless of whether it already exists
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
dest_table <- bq_table(dest_dataset, "Type2Diabetes")

# create a full table spec as follows
# then pass this to bq_table_upload
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("src_dateevent", "DATE"),
                              bq_field("t2d_status", "INT64")
))

# delete an existing table if necessary
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
# upload file_data to dest_table
bq_table_upload(dest_table, t2d, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')

nrow(t2d)
