==> code/4.Diabetes_Medications.ipynb <==
library(bigrquery)
library(tidyverse)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# Set your query
metformin <- "SELECT person_id, src_datemedicationstart, src_idmultilexdmd, src_nameofmedication, src_dateevent FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRPrimaryCareMedication` 
where src_nameofmedication LIKE '%metformin%' OR src_nameofmedication LIKE '%glucophage%' OR src_nameofmedication LIKE '%sukkarto%'"
thiazolid <- "SELECT person_id, src_datemedicationstart, src_idmultilexdmd, src_nameofmedication, src_dateevent FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRPrimaryCareMedication` 
where src_nameofmedication LIKE '%pioglitazone%' OR src_nameofmedication LIKE '%actos%' OR src_nameofmedication LIKE '%avandia%'"
gliptins <- "SELECT person_id, src_datemedicationstart, src_idmultilexdmd, src_nameofmedication, src_dateevent FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRPrimaryCareMedication` 
where src_nameofmedication LIKE '%gliptin%' OR src_nameofmedication LIKE '%vipidia%' OR src_nameofmedication LIKE '%trajenta%' OR src_nameofmedication LIKE '%onglyza%'
OR src_nameofmedication LIKE '%januvia%' OR src_nameofmedication LIKE '%galvus%'"
sulfon <- "SELECT person_id, src_datemedicationstart, src_idmultilexdmd, src_nameofmedication, src_dateevent FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRPrimaryCareMedication` 
where src_nameofmedication LIKE '%glimepiride%' OR src_nameofmedication LIKE '%amaryl%' OR src_nameofmedication LIKE '%gliclazide%' OR src_nameofmedication LIKE '%glipizide%'
OR src_nameofmedication LIKE '%glibenclamide%'"
glpra <- "SELECT person_id, src_datemedicationstart, src_idmultilexdmd, src_nameofmedication, src_dateevent FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRPrimaryCareMedication` 
where src_nameofmedication LIKE '%exenatide%' OR src_nameofmedication LIKE '%bydureon%' OR src_nameofmedication LIKE '%byetta%' OR src_nameofmedication LIKE '%dulaglutide%'
OR src_nameofmedication LIKE '%trulicity%' OR src_nameofmedication LIKE '%liraglutide%' OR src_nameofmedication LIKE '%victoza%' OR src_nameofmedication LIKE '%lixisenatide%'
 OR src_nameofmedication LIKE '%lyxumia%' OR src_nameofmedication LIKE '%semaglutide%' OR src_nameofmedication LIKE '%ozempic%'"
glucinhib <- "SELECT person_id, src_datemedicationstart, src_idmultilexdmd, src_nameofmedication, src_dateevent FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRPrimaryCareMedication` 
where src_nameofmedication LIKE '%flozin%' OR src_nameofmedication LIKE '%invokana%' OR src_nameofmedication LIKE '%forxiga%' OR src_nameofmedication LIKE '%jardiance%'"
combo <- "SELECT person_id, src_datemedicationstart, src_idmultilexdmd, src_nameofmedication, src_dateevent FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRPrimaryCareMedication` 
where src_nameofmedication LIKE '%alogliptin 12.5mg / metformin 1g tablets%' OR src_nameofmedication LIKE '%canagliflozin 50mg / metformin 1g tablets%' 
OR src_nameofmedication LIKE '%canagliflozin 50mg / metformin 850mg tablets%' OR src_nameofmedication LIKE '%dapagliflozin 5mg / metformin 1g tablets%'
OR src_nameofmedication LIKE '%dapagliflozin 5mg / metformin 850mg tablets%' OR src_nameofmedication LIKE '%empagliflozin 12.5mg / metformin 1g tablets%' 
OR src_nameofmedication LIKE '%empagliflozin 12.5mg / metformin 850mg tablets%' OR src_nameofmedication LIKE '%empagliflozin 5mg / metformin 1g tablets%'
OR src_nameofmedication LIKE '%empagliflozin 5mg / metformin 850mg tablets%' OR src_nameofmedication LIKE '%linagliptin 2.5mg / metformin 1g tablets%' 
OR src_nameofmedication LIKE '%linagliptin 2.5mg / metformin 850mg tablets%' OR src_nameofmedication LIKE '%metformin 1g / sitagliptin 50mg tablets%' 
OR src_nameofmedication LIKE '%pioglitazone 15mg / metformin 850mg tablets%' OR src_nameofmedication LIKE '%rosiglitazone 2mg / metformin 1g tablets%'
OR src_nameofmedication LIKE '%rosiglitazone 2mg / metformin 500mg tablets%' OR src_nameofmedication LIKE '%rosiglitazone 4mg / metformin 1g tablets%' 
OR src_nameofmedication LIKE '%saxagliptin 2.5mg / metformin 1g tablets%' OR src_nameofmedication LIKE '%saxagliptin 2.5mg / metformin 850mg tablets%' 
OR src_nameofmedication LIKE '%sitagliptin with metformin (roi) tablets 50mg + 850mg%' OR src_nameofmedication LIKE '%textual drug : metformin + pioglitazone tabs 850mg + 15mg%'
OR src_nameofmedication LIKE '%vildagliptin 50mg / metformin 1g tablets%' OR src_nameofmedication LIKE '%vildagliptin 50mg / metformin 850mg tablets%'"
insulin <- "SELECT person_id, src_datemedicationstart, src_idmultilexdmd, src_nameofmedication, src_dateevent FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRPrimaryCareMedication` 
where src_nameofmedication LIKE '%insulatard%' OR src_nameofmedication LIKE '%degludec%' OR src_nameofmedication LIKE '%tresiba%' OR src_nameofmedication LIKE '%detemir%'
OR src_nameofmedication LIKE '%glargine%' OR src_nameofmedication LIKE '%lantus%' OR src_nameofmedication LIKE '%toujeo%' OR src_nameofmedication LIKE '%abasaglar%'
OR src_nameofmedication LIKE '%humulin%' OR src_nameofmedication LIKE '%actrapid%' OR src_nameofmedication LIKE '%aspart%'
OR src_nameofmedication LIKE '%novorapid%' OR src_nameofmedication LIKE '%fiasp%'
OR src_nameofmedication LIKE '%lispro%' OR src_nameofmedication LIKE '%humalog%'
OR src_nameofmedication LIKE '%glulisine%' OR src_nameofmedication LIKE '%apidra%'
OR src_nameofmedication LIKE '%insuman%'"
drp_nt_drugs <- "SELECT person_id, src_datemedicationstart, src_idmultilexdmd, src_nameofmedication, src_dateevent FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRPrimaryCareMedication` 
where src_nameofmedication LIKE '%lactose%' OR src_nameofmedication LIKE '%becton%' OR src_nameofmedication LIKE '%menarini%' OR src_nameofmedication LIKE '%glucorx%'
OR src_nameofmedication LIKE '%gauge%' OR src_nameofmedication LIKE '%microfine%' OR src_nameofmedication LIKE '%micro-fine%' OR src_nameofmedication LIKE '%needles%'
OR src_nameofmedication LIKE '%clikstar%' OR src_nameofmedication LIKE '%reusable%' OR src_nameofmedication LIKE '%re-usable%'"

# Run the query; this returns a bq_table object that you can query further
metf <- bq_project_query(projectid, metformin)
thiaz <- bq_project_query(projectid, thiazolid)
glip <- bq_project_query(projectid, gliptins)
sulf <- bq_project_query(projectid, sulfon)
glp <- bq_project_query(projectid, glpra)
gluc <- bq_project_query(projectid, glucinhib)
comb <- bq_project_query(projectid, combo)
insu <- bq_project_query(projectid, insulin)
drop <- bq_project_query(projectid, drp_nt_drugs)

# Store data in a tibble
metformin <-bq_table_download(metf)
thiazolid <-bq_table_download(thiaz)
gliptins <-bq_table_download(glip)
sulfon <-bq_table_download(sulf)
glpra <-bq_table_download(glp)
glucinhib <-bq_table_download(gluc)
combo <- bq_table_download(comb)
insulin <- bq_table_download(insu)
drp_nt_drugs <- bq_table_download(drop)

# Filter all medication start dates after 2007-01-01
metformin <- metformin %>%
    mutate(src_datemedicationstart = as.Date(src_datemedicationstart)) %>%
    filter(src_datemedicationstart >= "2007-01-01" & src_datemedicationstart <= "2017-01-01")
thiazolid <- thiazolid %>%
    mutate(src_datemedicationstart = as.Date(src_datemedicationstart)) %>%
    filter(src_datemedicationstart >= "2007-01-01" & src_datemedicationstart <= "2017-01-01")
gliptins <- gliptins %>%
    mutate(src_datemedicationstart = as.Date(src_datemedicationstart)) %>%
    filter(src_datemedicationstart >= "2007-01-01" & src_datemedicationstart <= "2017-01-01")
sulfon <- sulfon %>%
    mutate(src_datemedicationstart = as.Date(src_datemedicationstart)) %>%
    filter(src_datemedicationstart >= "2007-01-01" & src_datemedicationstart <= "2017-01-01")
glpra <- glpra %>%
    mutate(src_datemedicationstart = as.Date(src_datemedicationstart)) %>%
    filter(src_datemedicationstart >= "2007-01-01" & src_datemedicationstart <= "2017-01-01")
glucinhib <- glucinhib %>%
    mutate(src_datemedicationstart = as.Date(src_datemedicationstart)) %>%
    filter(src_datemedicationstart >= "2007-01-01" & src_datemedicationstart <= "2017-01-01")
combo <- combo %>%
    mutate(src_datemedicationstart = as.Date(src_datemedicationstart)) %>%
    filter(src_datemedicationstart >= "2007-01-01" & src_datemedicationstart <= "2017-01-01")
insulin <- insulin %>%
    mutate(src_datemedicationstart = as.Date(src_datemedicationstart)) %>%
    filter(src_datemedicationstart >= "2007-01-01" & src_datemedicationstart <= "2017-01-01")

# Remove src_nameofmedication if in drp_nt_drugs
nrow(metformin)
metformin <- anti_join(metformin, drp_nt_drugs, by = "src_nameofmedication")
nrow(metformin)

nrow(thiazolid)
thiazolid <- anti_join(thiazolid, drp_nt_drugs, by = "src_nameofmedication")
nrow(thiazolid)

nrow(gliptins)
gliptins <- anti_join(gliptins, drp_nt_drugs, by = "src_nameofmedication")
nrow(gliptins)

nrow(sulfon)
sulfon <- anti_join(sulfon, drp_nt_drugs, by = "src_nameofmedication")
nrow(sulfon)

nrow(glpra)
glpra <- anti_join(glpra, drp_nt_drugs, by = "src_nameofmedication")
nrow(glpra)

nrow(glucinhib)
glucinhib <- anti_join(glucinhib, drp_nt_drugs, by = "src_nameofmedication")
nrow(glucinhib)

nrow(combo)
combo <- anti_join(combo, drp_nt_drugs, by = "src_nameofmedication")
nrow(combo)

nrow(insulin)
insulin <- anti_join(insulin, drp_nt_drugs, by = "src_nameofmedication")
nrow(insulin)

# Join with patient table to calculate age at event and gender
sql3 <- "SELECT person_id, birth_datetime, gender FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Patient_Denom`"
tb3 <- bq_project_query(projectid, sql3)
patient <- bq_table_download(tb3)
head(patient)

# Drop those where birth_datetime and gender = NA
patient <- patient %>%
    filter(!is.na(birth_datetime)) %>%
    filter(!is.na(gender))

# Join patient table to each drug table, calculate age at date medication start, keep earliest date for medication per person
# Metformin
pt_metf <- patient %>%
    right_join(metformin) %>%
    select(person_id, birth_datetime, gender, src_datemedicationstart) %>%
    mutate(age_metformin = floor(interval(birth_datetime, src_datemedicationstart) / years(1))) %>%
    group_by(person_id) %>%
    arrange(src_datemedicationstart) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number, with_ties = FALSE) %>%
    select(-event_number) %>%
    filter(!is.na(birth_datetime)) %>%
    rename(date_metformin = src_datemedicationstart) %>%
    mutate(has_metformin = 1) %>%
    select(-birth_datetime)

# Thiazolid
pt_thi <- patient %>%
    right_join(thiazolid) %>%
    select(person_id, birth_datetime, gender, src_datemedicationstart) %>%
    mutate(age_thiazolid = floor(interval(birth_datetime, src_datemedicationstart) / years(1))) %>%
    group_by(person_id) %>%
    arrange(src_datemedicationstart) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number, with_ties = FALSE) %>%
    select(-event_number) %>%
    filter(!is.na(birth_datetime)) %>%
    rename(date_thiazolid = src_datemedicationstart) %>%
    mutate(has_thiazolid = 1) %>%
    select(-birth_datetime)

# Gliptins
pt_glip <- patient %>%
    right_join(gliptins) %>%
    select(person_id, birth_datetime, gender, src_datemedicationstart) %>%
    mutate(age_gliptins = floor(interval(birth_datetime, src_datemedicationstart) / years(1))) %>%
    group_by(person_id) %>%
    arrange(src_datemedicationstart) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number, with_ties = FALSE) %>%
    select(-event_number) %>%
    filter(!is.na(birth_datetime)) %>%
    rename(date_gliptins = src_datemedicationstart) %>%
    mutate(has_gliptins = 1) %>%
    select(-birth_datetime)

# Sulfon
pt_sulf <- patient %>%
    right_join(sulfon) %>%
    select(person_id, birth_datetime, gender, src_datemedicationstart) %>%
    mutate(age_sulfon = floor(interval(birth_datetime, src_datemedicationstart) / years(1))) %>%
    group_by(person_id) %>%
    arrange(src_datemedicationstart) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number, with_ties = FALSE) %>%
    select(-event_number) %>%
    filter(!is.na(birth_datetime)) %>%
    rename(date_sulfon = src_datemedicationstart) %>%
    mutate(has_sulfon = 1) %>%
    select(-birth_datetime)

# Glpra
pt_glpra <- patient %>%
    right_join(glpra) %>%
    select(person_id, birth_datetime, gender, src_datemedicationstart) %>%
    mutate(age_glpra = floor(interval(birth_datetime, src_datemedicationstart) / years(1))) %>%
    group_by(person_id) %>%
    arrange(src_datemedicationstart) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number, with_ties = FALSE) %>%
    select(-event_number) %>%
    filter(!is.na(birth_datetime)) %>%
    rename(date_glpra = src_datemedicationstart) %>%
    mutate(has_glpra = 1) %>%
    select(-birth_datetime)

# Glucinhib
pt_glu <- patient %>%
    right_join(glucinhib) %>%
    select(person_id, birth_datetime, gender, src_datemedicationstart) %>%
    mutate(age_glucinhib = floor(interval(birth_datetime, src_datemedicationstart) / years(1))) %>%
    group_by(person_id) %>%
    arrange(src_datemedicationstart) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number, with_ties = FALSE) %>%
    select(-event_number) %>%
    filter(!is.na(birth_datetime)) %>%
    rename(date_glucinhib = src_datemedicationstart) %>%
    mutate(has_glucinhib = 1) %>%
    select(-birth_datetime)

# Combo
pt_comb <- patient %>%
    right_join(combo) %>%
    select(person_id, birth_datetime, gender, src_datemedicationstart) %>%
    mutate(age_combo = floor(interval(birth_datetime, src_datemedicationstart) / years(1))) %>%
    group_by(person_id) %>%
    arrange(src_datemedicationstart) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number, with_ties = FALSE) %>%
    select(-event_number) %>%
    filter(!is.na(birth_datetime)) %>%
    rename(date_combo = src_datemedicationstart) %>%
    mutate(has_combo = 1) %>%
    select(-birth_datetime)

# Insulin
pt_insu <- patient %>%
    right_join(insulin) %>%
    select(person_id, birth_datetime, gender, src_datemedicationstart) %>%
    mutate(age_insulin = floor(interval(birth_datetime, src_datemedicationstart) / years(1))) %>%
    group_by(person_id) %>%
    arrange(src_datemedicationstart) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number, with_ties = FALSE) %>%
    select(-event_number) %>%
    filter(!is.na(birth_datetime)) %>%
    rename(date_insulin = src_datemedicationstart) %>%
    mutate(has_insulin = 1) %>%
    select(-birth_datetime)

nrow(pt_metf)
nrow(pt_thi)
nrow(pt_glip)
nrow(pt_sulf)
nrow(pt_glpra)
nrow(pt_glu)
nrow(pt_comb)
nrow(pt_insu)

# Join all tables by person_id
meds <- pt_metf %>%
    left_join(pt_thi, by = c("person_id", "gender")) %>%
    left_join(pt_glip, by = c("person_id", "gender")) %>%
    left_join(pt_sulf, by = c("person_id", "gender")) %>%
    left_join(pt_glpra, by = c("person_id", "gender")) %>%
    left_join(pt_glu, by = c("person_id", "gender")) %>%
    left_join(pt_comb, by = c("person_id", "gender")) %>%
    left_join(pt_insu, by = c("person_id", "gender"))

# Add variable for earliest date
meds <- meds %>%
    mutate(date_earliest_med = pmin(date_metformin, date_thiazolid, date_gliptins, date_sulfon, date_glpra, date_glucinhib, date_combo, date_insulin, na.rm = TRUE))

# add indicator for metformin i.e. if taking metformin and another drug (1 = metformin + another drug, 2 = only metformin), if male or female > 50 years 
meds <- meds %>%
    mutate(met_plus = case_when(
        !is.na(has_metformin) & (has_thiazolid == 1 | has_gliptins == 1 | has_sulfon == 1 | has_glpra == 1 | has_glucinhib == 1 | has_combo == 1 | has_insulin == 1) ~ 1,
        !is.na(has_metformin) & is.na(has_thiazolid) & is.na(has_gliptins) & is.na(has_sulfon) & is.na(has_glpra) & is.na(has_glucinhib) & is.na(has_combo) & is.na(has_insulin) ~ 2)) %>%
    mutate(met_diab = ifelse((met_plus == 2 & gender == 1) | (met_plus == 2 & gender == 2 & age_metformin >= 50),1,0)) %>%
    mutate(met_fe_lsthn_50 = ifelse(met_plus == 2 & gender == 2 & age_metformin < 50,1,0)) %>%
    mutate(met_earliest = ifelse(date_metformin == date_earliest_med,1,0))

colnames(meds)

# Save dataset
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
dest_table <- bq_table(dest_dataset, "Diabetic_Meds")

# create a full table spec as follows
# then pass this to bq_table_upload
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("gender", "INT64"),
                              bq_field("date_metformin", "DATE"),
                              bq_field("age_metformin", "INT64"),
                              bq_field("has_metformin", "INT64"),
                              bq_field("date_thiazolid", "DATE"),
                              bq_field("age_thiazolid", "INT64"),
                              bq_field("has_thiazolid", "INT64"),
                              bq_field("date_gliptins", "DATE"),
                              bq_field("age_gliptins", "INT64"),
                              bq_field("has_gliptins", "INT64"),
                              bq_field("date_sulfon", "DATE"),
                              bq_field("age_sulfon", "INT64"),
                              bq_field("has_sulfon", "INT64"),
                              bq_field("date_glpra", "DATE"),
                              bq_field("age_glpra", "INT64"),
                              bq_field("has_glpra", "INT64"),
                              bq_field("date_glucinhib", "DATE"),
                              bq_field("age_glucinhib", "INT64"),
                              bq_field("has_glucinhib", "INT64"),
                              bq_field("date_combo", "DATE"),
                              bq_field("age_combo", "INT64"),
                              bq_field("has_combo", "INT64"),
                              bq_field("date_insulin", "DATE"),
                              bq_field("age_insulin", "INT64"),
                              bq_field("has_insulin", "INT64"),
                              bq_field("date_earliest_med", "DATE"),
                              bq_field("met_plus", "INT64"),
                              bq_field("met_diab", "INT64"),
                              bq_field("met_earliest", "INT64"),
                              bq_field("met_fe_lsthn_50", "INT64")))

# delete an existing table if necessary
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
# upload file_data to dest_table
bq_table_upload(dest_table, meds, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')
