==> code/5.Joining_T1D_T2D_Meds.ipynb <==
library(bigrquery)
library(tidyverse)
library(lubridate)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# Set your query
sql1 <- "SELECT * FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Type1Diabetes`"
sql2 <- "SELECT * FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Type2Diabetes`"
sql3 <- "SELECT * FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Diabetic_Meds`"

# Run the query; this returns a bq_table object that you can query further
tb1 <- bq_project_query(projectid, sql1)
tb2 <- bq_project_query(projectid, sql2)
tb3 <- bq_project_query(projectid, sql3)

# Store data in a tibble
t1d <-bq_table_download(tb1)
t2d <-bq_table_download(tb2)
meds <-bq_table_download(tb3)

# join t1d and t2d tables
diabetes <- full_join(t1d, t2d, by = "person_id")
head(diabetes)
rm(t1d, t2d)

diabetes <- diabetes %>%
    mutate(t1d_status = ifelse(is.na(t1d_status),0,1)) %>%
    mutate(t2d_status = ifelse(is.na(t2d_status),0,1)) %>%
    rename(date_t1d_diag = src_dateevent.x) %>%
    rename(date_t2d_diag = src_dateevent.y) %>%
    mutate(any_diabetes = ifelse(t1d_status == 1 | t2d_status == 1,1,0))

head(diabetes)

nrow(diabetes)
print("type 1 and type 2 diabetes")
table(diabetes$t1d_status, diabetes$t2d_status)
print("type 1 diabetes")
table(diabetes$t1d_status)
print("type 2 diabetes")
table(diabetes$t2d_status)
sum(duplicated(diabetes))

# Join patient table to diabetes patients
sql3 <- "SELECT * FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Patient_Denom`"
tb3 <- bq_project_query(projectid, sql3)
patient <- bq_table_download(tb3)
head(patient)

sum(is.na(patient$birth_datetime))
nrow(patient)

diabetes_patient <- inner_join(patient, diabetes, by = "person_id")
head(diabetes_patient)

# calculate age at diabetes
diabetes_patient$age_t1d_diag <- interval(diabetes_patient$birth_datetime, diabetes_patient$date_t1d_diag) / years(1)
diabetes_patient$age_t1d_diag <- floor(diabetes_patient$age_t1d_diag)
diabetes_patient$age_t2d_diag <- interval(diabetes_patient$birth_datetime, diabetes_patient$date_t2d_diag) / years(1)
diabetes_patient$age_t2d_diag <- floor(diabetes_patient$age_t2d_diag)
head(diabetes_patient)

# age at earliest diabetes 
diabetes_patient <- diabetes_patient %>%
    mutate(age_earliest_diabetes = case_when(
    (age_t1d_diag <= age_t2d_diag) | is.na(age_t2d_diag) ~ age_t1d_diag,
    (age_t1d_diag > age_t2d_diag) | is.na(age_t1d_diag) ~ age_t2d_diag)) %>%
select(-gender, -ethnicity, -imd_decile)
sum(is.na(diabetes_patient$birth_datetime))
nrow(diabetes_patient)

table(diabetes_patient$t2d_status)

(15228/43568)*100

# add date at earliest diabetes
diabetes_patient <- diabetes_patient %>%
    mutate(date_earliest_diabetes = case_when(
    (date_t1d_diag <= date_t2d_diag) | is.na(date_t2d_diag) ~ date_t1d_diag,
    (date_t1d_diag > date_t2d_diag) | is.na(date_t1d_diag) ~ date_t2d_diag,
    is.na(date_t1d_diag) & is.na(date_t2d_diag) ~ NA))
head(diabetes_patient)

nrow(diabetes_patient)
table(diabetes_patient$t1d_status)
table(diabetes_patient$t2d_status)
table(diabetes_patient$t1d_status, diabetes_patient$t2d_status)

# Join medication data to diabetes_patient
dm <- full_join(diabetes_patient, meds, by = "person_id")

nrow(dm)

# Remove females who are less than 50 who are on metformin but don't have a diabetes diagnostic code
dm <- dm %>%
    mutate(diab_date_present = ifelse(!is.na(date_earliest_diabetes),1,0)) %>% # 1 = diabetes date is present i.e. patient has diabetes diagnostic code
    mutate(metformin_pls_5yrs = (date_metformin %m+% years(5))) %>% # calculate 5 years from metformin prescription
    mutate(diab_met_diff = abs(floor(difftime(date_earliest_diabetes, date_metformin, units = c("days"))))) %>% # absolute number of days between diabetes diagnosis date and metformin date - if >1826 days, drop
    mutate(diab_met_diff = as.integer(diab_met_diff))

dm %>%
select(date_earliest_diabetes, diab_date_present, metformin_pls_5yrs, diab_met_diff, met_fe_lsthn_50, met_plus, gender) %>%
filter(!is.na(diab_met_diff)) %>%
arrange(desc(diab_met_diff)) 

colnames(dm)

# diabetes diagnosis from medications, choose earliest diabetes diagnosis date between earliest date and earliest med, calculate age at earliest diabetes diag
dm <- dm %>%
    mutate(diabetes_meds = ifelse(is.na(date_earliest_diabetes) & (has_metformin == 1 | has_thiazolid == 1 | has_gliptins == 1 | has_sulfon == 1 | has_glpra == 1 |
                                                                  has_glucinhib == 1 | has_combo == 1 | has_insulin == 1),1,0)) %>%
    mutate(earliest_diabetes_diag = case_when(
    (date_earliest_diabetes <= date_earliest_med) | is.na(date_earliest_med) ~ date_earliest_diabetes,
    (date_earliest_diabetes > date_earliest_med) | is.na(date_earliest_diabetes) ~ date_earliest_med)) %>%
    mutate(age_earliest_diab = floor(interval(birth_datetime, earliest_diabetes_diag) / years(1))) %>%
    # time between diabetes and death
    mutate(diab_death_time = abs(difftime(earliest_diabetes_diag, death_datetime, units = "days")),
          diab_death_time = as.integer(diab_death_time))

dm <- dm %>% select(person_id, date_t1d_diag, age_t1d_diag, t1d_status, date_t2d_diag, age_t2d_diag, t2d_status, any_diabetes, diabetes_meds, age_earliest_diab, earliest_diabetes_diag, diab_death_time)

nrow(dm)
table(dm$t1d_status)
table(dm$t2d_status)
table(dm$t1d_status, dm$t2d_status)
table(dm$any_diabetes)

colnames(dm)

# set destination table - do this regardless of whether it already exists
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
dest_table <- bq_table(dest_dataset, "Diabetes")

# create a full table spec as follows
# then pass this to bq_table_upload
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("age_t1d_diag", "INT64"),
                              bq_field("date_t1d_diag", "DATE"),
                              bq_field("t1d_status", "INT64"),
                              bq_field("age_t2d_diag", "INT64"),
                              bq_field("date_t2d_diag", "DATE"),
                              bq_field("t2d_status", "INT64"),
                              bq_field("age_earliest_diab", "INT64"),
                              bq_field("earliest_diabetes_diag", "DATE"),
                              bq_field("any_diabetes", "INT64"),
                              bq_field("diab_death_time", "INT64"),
                              bq_field("diabetes_meds", "INT64")
))

# delete an existing table if necessary
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
# upload file_data to dest_table
bq_table_upload(dest_table, dm, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')

sum(is.na(diabetes_patient$gender))

nrow(diabetes_patient)

43568-15244
