==> code/6.CVD_PrimaryCare.ipynb <==
library(bigrquery)
library(tidyverse)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# Angina: Primary care between 2002-2022
sql1 <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` 
WHERE src_dateevent BETWEEN '2002-01-01' AND '2022-12-31' AND (src_ctv3code LIKE '%G311.%' OR src_ctv3code LIKE '%G31y0%' OR src_ctv3code LIKE '%G33..%' OR src_ctv3code LIKE '%G33..%'
OR src_ctv3code LIKE '%G33..%' OR src_ctv3code LIKE '%G33..%' OR src_ctv3code LIKE '%G33..%' OR src_ctv3code LIKE '%G33..%' OR src_ctv3code LIKE '%G33..%' OR src_ctv3code LIKE '%G33..%'
OR src_ctv3code LIKE '%G33..%' OR src_ctv3code LIKE '%G33..%' OR src_ctv3code LIKE '%G33..%' OR src_ctv3code LIKE '%G3300%' OR src_ctv3code LIKE '%G331.%' OR src_ctv3code LIKE '%G33z3%')"
tb1 <- bq_project_query(projectid, sql1)
angina <-bq_table_download(tb1) # store data in tibble

# Cardiomyopathy: Primary care between 2002-2022
sql2 <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` 
WHERE src_dateevent BETWEEN '2002-01-01' AND '2022-12-31' AND (src_ctv3code LIKE '%C3733%' OR src_ctv3code LIKE '%C3733%' OR src_ctv3code LIKE '%G343.%' OR src_ctv3code LIKE '%G55..%'
OR src_ctv3code LIKE '%G55..%' OR src_ctv3code LIKE '%G55..%' OR src_ctv3code LIKE '%G55..%' OR src_ctv3code LIKE '%G55..%' OR src_ctv3code LIKE '%G55..%' OR src_ctv3code LIKE '%G55..%' 
OR src_ctv3code LIKE '%G55..%' OR src_ctv3code LIKE '%G55..%' OR src_ctv3code LIKE '%G550.%' OR src_ctv3code LIKE '%G551.%' OR src_ctv3code LIKE '%G554.%' OR src_ctv3code LIKE '%G5541%'
OR src_ctv3code LIKE '%G5543%' OR src_ctv3code LIKE '%G5544%' OR src_ctv3code LIKE '%G556.%' OR src_ctv3code LIKE '%G556.%' OR src_ctv3code LIKE '%G556.%' OR src_ctv3code LIKE '%G556.%'
OR src_ctv3code LIKE '%G557.%' OR src_ctv3code LIKE '%G557.%' OR src_ctv3code LIKE '%G5572%' OR src_ctv3code LIKE '%G5574%' OR src_ctv3code LIKE '%G5574%' OR src_ctv3code LIKE '%G558.%' 
OR src_ctv3code LIKE '%G5580%' OR src_ctv3code LIKE '%G5581%' OR src_ctv3code LIKE '%G55y0%' OR src_ctv3code LIKE '%G5y1.%' OR src_ctv3code LIKE '%Gyu5Q%' OR src_ctv3code LIKE '%Gyu5S%'
OR src_ctv3code LIKE '%X708p%' OR src_ctv3code LIKE '%X776w%' OR src_ctv3code LIKE '%XaI9b%')"
tb2 <- bq_project_query(projectid, sql2)
cardiomyopathy <-bq_table_download(tb2) # store data in tibble

# Heart Failure: Primary care between 2002-2022
sql3 <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` 
WHERE src_dateevent BETWEEN '2002-01-01' AND '2022-12-31' AND (src_ctv3code LIKE '%G1yz1%' OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%'
OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%'
OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G58..%' OR src_ctv3code LIKE '%G580.%'
OR src_ctv3code LIKE '%G5802%' OR src_ctv3code LIKE '%G5810%' OR src_ctv3code LIKE '%G5810%' OR src_ctv3code LIKE '%G5810%' OR src_ctv3code LIKE '%G5810%' OR src_ctv3code LIKE '%G5y4.%')"
tb3 <- bq_project_query(projectid, sql3)
heart_failure <-bq_table_download(tb3) # store data in tibble

# Hypertension: Primary care between 2002-2022
sql4 <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` 
WHERE src_dateevent BETWEEN '2002-01-01' AND '2022-12-31' AND (src_ctv3code LIKE '%61462%' OR src_ctv3code LIKE '%G2...%' OR src_ctv3code LIKE '%G20..%' OR src_ctv3code LIKE '%G20..%'
OR src_ctv3code LIKE '%G20..%' OR src_ctv3code LIKE '%G20..%' OR src_ctv3code LIKE '%G20..%' OR src_ctv3code LIKE '%G20..%' OR src_ctv3code LIKE '%G20..%' OR src_ctv3code LIKE '%G20..%'
OR src_ctv3code LIKE '%G20..%' OR src_ctv3code LIKE '%G20..%' OR src_ctv3code LIKE '%G20..%' OR src_ctv3code LIKE '%G20..%' OR src_ctv3code LIKE '%G200.%' OR src_ctv3code LIKE '%G200.%'
OR src_ctv3code LIKE '%G200.%' OR src_ctv3code LIKE '%G22z.%' OR src_ctv3code LIKE '%G24..%' OR src_ctv3code LIKE '%G24..%' OR src_ctv3code LIKE '%G24..%' OR src_ctv3code LIKE '%G2400%'
OR src_ctv3code LIKE '%G244.%' OR src_ctv3code LIKE '%Gyu21%' OR src_ctv3code LIKE '%L12..%' OR src_ctv3code LIKE '%L120.%' OR src_ctv3code LIKE '%L121.%' OR src_ctv3code LIKE '%L122.%'
OR src_ctv3code LIKE '%L125.%' OR src_ctv3code LIKE '%L1250%' OR src_ctv3code LIKE '%L126.%' OR src_ctv3code LIKE '%L1261%' OR src_ctv3code LIKE '%L1263%' OR src_ctv3code LIKE '%X40By%'
OR src_ctv3code LIKE '%X40Bz%' OR src_ctv3code LIKE '%X40C2%' OR src_ctv3code LIKE '%X40C3%' OR src_ctv3code LIKE '%XaBE0%')"
tb4 <- bq_project_query(projectid, sql4)
hypertension <-bq_table_download(tb4) # store data in tibble

# Ischaemic Heart Disease: Primary care between 2002-2022
sql5 <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` 
WHERE src_dateevent BETWEEN '2002-01-01' AND '2022-12-31' AND (src_ctv3code LIKE '%G30y0%' OR src_ctv3code LIKE '%G30y1%' OR src_ctv3code LIKE '%G30y2%' OR src_ctv3code LIKE '%G31..%'
OR src_ctv3code LIKE '%G34..%' OR src_ctv3code LIKE '%G3400%' OR src_ctv3code LIKE '%G341.%' OR src_ctv3code LIKE '%G3413%' OR src_ctv3code LIKE '%G34y1%' OR src_ctv3code LIKE '%G363.%'
OR src_ctv3code LIKE '%X200C%' OR src_ctv3code LIKE '%X200C%' OR src_ctv3code LIKE '%X200d%')"
tb5 <- bq_project_query(projectid, sql5)
ihd <-bq_table_download(tb5) # store data in tibble

# Myocardial Infarction: Primary care between 2002-2022
sql6 <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` 
WHERE src_dateevent BETWEEN '2002-01-01' AND '2022-12-31' AND (src_ctv3code LIKE '%G300.%' OR src_ctv3code LIKE '%G300.%' OR src_ctv3code LIKE '%G302.%' OR src_ctv3code LIKE '%G302.%'
OR src_ctv3code LIKE '%G35..%' OR src_ctv3code LIKE '%G353.%' OR src_ctv3code LIKE '%Gyu34%' OR src_ctv3code LIKE '%X200W%' OR src_ctv3code LIKE '%XE0Uh%' OR src_ctv3code LIKE '%XE0Uh%')"
tb6 <- bq_project_query(projectid, sql6)
mi <-bq_table_download(tb6) # store data in tibble

# Peripheral Vascular Disease: Primary care between 2002-2022
sql7 <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` 
WHERE src_dateevent BETWEEN '2002-01-01' AND '2022-12-31' AND (src_ctv3code LIKE '%G73y.%' OR src_ctv3code LIKE '%G73y0%' OR src_ctv3code LIKE '%G73y0%' OR src_ctv3code LIKE '%G73y0%'
OR src_ctv3code LIKE '%G73y0%' OR src_ctv3code LIKE '%G73y0%' OR src_ctv3code LIKE '%G73y0%' OR src_ctv3code LIKE '%G73y0%' OR src_ctv3code LIKE '%G73y0%' OR src_ctv3code LIKE '%G73y0%' OR src_ctv3code LIKE '%Xa0lV%')"
tb7 <- bq_project_query(projectid, sql7)
pvd <-bq_table_download(tb7) # store data in tibble

# Stroke: Primary care between 2002-2022
sql8 <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` 
WHERE src_dateevent BETWEEN '2002-01-01' AND '2022-12-31' AND (src_ctv3code LIKE '%XaBM5%' OR src_ctv3code LIKE '%XaBM5%' OR src_ctv3code LIKE '%XaBM5%' OR src_ctv3code LIKE '%Gyu62%'
OR src_ctv3code LIKE '%G65..%' OR src_ctv3code LIKE '%XE0VK%' OR src_ctv3code LIKE '%X00DU%' OR src_ctv3code LIKE '%X00DW%' OR src_ctv3code LIKE '%Xa00T%' OR src_ctv3code LIKE '%Xa01k%'
OR src_ctv3code LIKE '%G610.%' OR src_ctv3code LIKE '%Xa00M%' OR src_ctv3code LIKE '%Xa00M%' OR src_ctv3code LIKE '%G662.%' OR src_ctv3code LIKE '%G662.%' OR src_ctv3code LIKE '%G613.%'
OR src_ctv3code LIKE '%G613.%' OR src_ctv3code LIKE '%G613.%' OR src_ctv3code LIKE '%G63y1%' OR src_ctv3code LIKE '%G601.%' OR src_ctv3code LIKE '%XaJgQ%' OR src_ctv3code LIKE '%14AB.%'
OR src_ctv3code LIKE '%Gyu6F%' OR src_ctv3code LIKE '%G668.%' OR src_ctv3code LIKE '%G668.%' OR src_ctv3code LIKE '%G668.%' OR src_ctv3code LIKE '%G668.%' OR src_ctv3code LIKE '%G668.%'
OR src_ctv3code LIKE '%G668.%' OR src_ctv3code LIKE '%Xa00K%' OR src_ctv3code LIKE '%G641.%' OR src_ctv3code LIKE '%G641.%' OR src_ctv3code LIKE '%G641.%' OR src_ctv3code LIKE '%G641.%'
OR src_ctv3code LIKE '%G641.%' OR src_ctv3code LIKE '%XaQbK%' OR src_ctv3code LIKE '%X00DB%' OR src_ctv3code LIKE '%Xa1uW%' OR src_ctv3code LIKE '%Xa1uW%' OR src_ctv3code LIKE '%Xa1uW%'
OR src_ctv3code LIKE '%Xa1uW%' OR src_ctv3code LIKE '%Gyu66%' OR src_ctv3code LIKE '%G654.%' OR src_ctv3code LIKE '%Xa01h%' OR src_ctv3code LIKE '%G660.%' OR src_ctv3code LIKE '%G660.%'
OR src_ctv3code LIKE '%G615.%' OR src_ctv3code LIKE '%G63y0%' OR src_ctv3code LIKE '%Gyu6G%' OR src_ctv3code LIKE '%XaQbM%' OR src_ctv3code LIKE '%F4236%' OR src_ctv3code LIKE '%G618.%'
OR src_ctv3code LIKE '%Fyu55%' OR src_ctv3code LIKE '%Gyu60%' OR src_ctv3code LIKE '%G600.%' OR src_ctv3code LIKE '%XaBL3%' OR src_ctv3code LIKE '%G617.%' OR src_ctv3code LIKE '%X00E2%'
OR src_ctv3code LIKE '%X00E2%' OR src_ctv3code LIKE '%G65z0%' OR src_ctv3code LIKE '%G6760%' OR src_ctv3code LIKE '%Xa01l%' OR src_ctv3code LIKE '%G661.%' OR src_ctv3code LIKE '%G661.%'
OR src_ctv3code LIKE '%Fyu56%' OR src_ctv3code LIKE '%Gyu64%' OR src_ctv3code LIKE '%Gyu61%' OR src_ctv3code LIKE '%G606.%' OR src_ctv3code LIKE '%XE0VF%' OR src_ctv3code LIKE '%XE0VF%'
OR src_ctv3code LIKE '%G664.%' OR src_ctv3code LIKE '%Xa01j%' OR src_ctv3code LIKE '%G6...%' OR src_ctv3code LIKE '%G6...%' OR src_ctv3code LIKE '%G6...%' OR src_ctv3code LIKE '%G6...%'
OR src_ctv3code LIKE '%X204A%' OR src_ctv3code LIKE '%Xa3fW%' OR src_ctv3code LIKE '%XE0Wy%' OR src_ctv3code LIKE '%XE0Wy%' OR src_ctv3code LIKE '%XM0rV%' OR src_ctv3code LIKE '%XM0rV%' OR src_ctv3code LIKE '%Gyu65%')"
tb8 <- bq_project_query(projectid, sql8)
stroke <-bq_table_download(tb8) # store data in tibble

# Valve Disease: Primary care between 2002-2022
sql9 <- "SELECT person_id, src_dateevent, src_ctv3code, src_ctv3text FROM `yhcr-prd-phm-bia-core.CB_FDM_PrimaryCare_v5.tbl_SRCode` 
WHERE src_dateevent BETWEEN '2002-01-01' AND '2022-12-31' AND (src_ctv3code LIKE '%G11..%' OR src_ctv3code LIKE '%G11..%' OR src_ctv3code LIKE '%G11..%' OR src_ctv3code LIKE '%G11..%'
OR src_ctv3code LIKE '%G11..%' OR src_ctv3code LIKE '%G11..%' OR src_ctv3code LIKE '%G110.%' OR src_ctv3code LIKE '%G111.%' OR src_ctv3code LIKE '%G111.%' OR src_ctv3code LIKE '%G111.%'
OR src_ctv3code LIKE '%G113.%' OR src_ctv3code LIKE '%G12..%' OR src_ctv3code LIKE '%G12..%' OR src_ctv3code LIKE '%G120.%' OR src_ctv3code LIKE '%G120.%' OR src_ctv3code LIKE '%G120.%'
OR src_ctv3code LIKE '%G121.%' OR src_ctv3code LIKE '%G140.%' OR src_ctv3code LIKE '%G1400%' OR src_ctv3code LIKE '%G1400%' OR src_ctv3code LIKE '%G1401%' OR src_ctv3code LIKE '%G140z%'
OR src_ctv3code LIKE '%G1410%' OR src_ctv3code LIKE '%G364.%' OR src_ctv3code LIKE '%G365.%' OR src_ctv3code LIKE '%G540.%' OR src_ctv3code LIKE '%G5400%' OR src_ctv3code LIKE '%G540z%'
OR src_ctv3code LIKE '%G541.%' OR src_ctv3code LIKE '%G541.%' OR src_ctv3code LIKE '%G541.%' OR src_ctv3code LIKE '%G5410%' OR src_ctv3code LIKE '%G5410%' OR src_ctv3code LIKE '%G5411%'
OR src_ctv3code LIKE '%G5411%' OR src_ctv3code LIKE '%G5411%' OR src_ctv3code LIKE '%G542.%' OR src_ctv3code LIKE '%G542.%' OR src_ctv3code LIKE '%G5420%' OR src_ctv3code LIKE '%G5421%'
OR src_ctv3code LIKE '%G5422%' OR src_ctv3code LIKE '%G542z%' OR src_ctv3code LIKE '%G543.%' OR src_ctv3code LIKE '%G543.%' OR src_ctv3code LIKE '%G543.%' OR src_ctv3code LIKE '%G543.%'
OR src_ctv3code LIKE '%G543.%' OR src_ctv3code LIKE '%G543.%' OR src_ctv3code LIKE '%G5430%' OR src_ctv3code LIKE '%G5431%' OR src_ctv3code LIKE '%G5434%' OR src_ctv3code LIKE '%G544.%'
OR src_ctv3code LIKE '%G544.%' OR src_ctv3code LIKE '%G544.%' OR src_ctv3code LIKE '%G54z.%' OR src_ctv3code LIKE '%G54z4%' OR src_ctv3code LIKE '%G5y6.%' OR src_ctv3code LIKE '%Gyu10%'
OR src_ctv3code LIKE '%Gyu11%' OR src_ctv3code LIKE '%Gyu55%' OR src_ctv3code LIKE '%Gyu56%' OR src_ctv3code LIKE '%Gyu57%' OR src_ctv3code LIKE '%Gyu58%' OR src_ctv3code LIKE '%Gyu59%'
OR src_ctv3code LIKE '%Gyu5B%' OR src_ctv3code LIKE '%P601.%' OR src_ctv3code LIKE '%P61..%' OR src_ctv3code LIKE '%P62..%' OR src_ctv3code LIKE '%P65..%' OR src_ctv3code LIKE '%P652.%'
OR src_ctv3code LIKE '%P6y0.%' OR src_ctv3code LIKE '%P7224%' OR src_ctv3code LIKE '%Pyu23%' OR src_ctv3code LIKE '%Pyu2G%' OR src_ctv3code LIKE '%X200s%' OR src_ctv3code LIKE '%X200u%'
OR src_ctv3code LIKE '%X2010%' OR src_ctv3code LIKE '%X2010%' OR src_ctv3code LIKE '%X2019%' OR src_ctv3code LIKE '%X201B%' OR src_ctv3code LIKE '%X201C%' OR src_ctv3code LIKE '%X777Y%'
OR src_ctv3code LIKE '%X778h%' OR src_ctv3code LIKE '%X778h%' OR src_ctv3code LIKE '%X778h%' OR src_ctv3code LIKE '%X778h%' OR src_ctv3code LIKE '%X77vL%' OR src_ctv3code LIKE '%X77zY%'
OR src_ctv3code LIKE '%XE0Ux%' OR src_ctv3code LIKE '%XE2bE%' OR src_ctv3code LIKE '%XM0rP%' OR src_ctv3code LIKE '%XSDVN%' OR src_ctv3code LIKE '%Xa7rt%')"
tb9 <- bq_project_query(projectid, sql9)
valve_disease <-bq_table_download(tb9) # store data in tibble

print("Angina")
head(angina)

print("Cardiomyopathy")
head(cardiomyopathy)

print("Heart Failure")
head(heart_failure)

print("Hypertension")
head(hypertension)

print("IHD")
head(ihd)

print("MI")
head(mi)

print("PVD")
head(pvd)

print("Stroke")
head(stroke)

print("Valve Disease")
head(valve_disease)

angina <- angina[!duplicated(angina),]
cardiomyopathy <- cardiomyopathy[!duplicated(cardiomyopathy),]
heart_failure <- heart_failure[!duplicated(heart_failure),]
hypertension <- hypertension[!duplicated(hypertension),]
ihd <- ihd[!duplicated(ihd),]
mi <- mi[!duplicated(mi),]
pvd <- pvd[!duplicated(pvd),]
stroke <- stroke[!duplicated(stroke),]
valve_disease <- valve_disease[!duplicated(valve_disease),]

nrow(angina)
nrow(cardiomyopathy)
nrow(heart_failure)
nrow(hypertension)
nrow(ihd)
nrow(mi)
nrow(pvd)
nrow(stroke)
nrow(valve_disease)

# add indicator for each CVD, rename date of event, remove ctv3 code and text columns, choose earliest event of CVD per person
# Angina
angina <- angina %>%
    mutate(has_angina = 1) %>%
    rename(date_angina_diagnosis = src_dateevent) %>%
    select(-src_ctv3code, -src_ctv3text) %>%
    group_by(person_id) %>%
    arrange(date_angina_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Cardiomyopathy
cardiomyopathy <- cardiomyopathy %>%
    mutate(has_cardiomyopathy = 1) %>%
    rename(date_cardiomyopathy_diagnosis = src_dateevent) %>%
    select(-src_ctv3code, -src_ctv3text) %>%
    group_by(person_id) %>%
    arrange(date_cardiomyopathy_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Heart Failure
heart_failure <- heart_failure %>%
    mutate(has_heart_failure = 1) %>%
    rename(date_heart_failure_diagnosis = src_dateevent) %>%
    select(-src_ctv3code, -src_ctv3text) %>%
    group_by(person_id) %>%
    arrange(date_heart_failure_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Hypertension
hypertension <- hypertension %>%
    mutate(has_hypertension = 1) %>%
    rename(date_hypertension_diagnosis = src_dateevent) %>%
    select(-src_ctv3code, -src_ctv3text) %>%
    group_by(person_id) %>%
    arrange(date_hypertension_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Ischaemic Heart Disease
ihd <- ihd %>%
    mutate(has_ihd = 1) %>%
    rename(date_ihd_diagnosis = src_dateevent) %>%
    select(-src_ctv3code, -src_ctv3text) %>%
    group_by(person_id) %>%
    arrange(date_ihd_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Myocardial Infarction
mi <- mi %>%
    mutate(has_mi = 1) %>%
    rename(date_mi_diagnosis = src_dateevent) %>%
    select(-src_ctv3code, -src_ctv3text) %>%
    group_by(person_id) %>%
    arrange(date_mi_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Peripheral Vascular Disease
pvd <- pvd %>%
    mutate(has_pvd = 1) %>%
    rename(date_pvd_diagnosis = src_dateevent) %>%
    select(-src_ctv3code, -src_ctv3text) %>%
    group_by(person_id) %>%
    arrange(date_pvd_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Stroke
stroke <- stroke %>%
    mutate(has_stroke = 1) %>%
    rename(date_stroke_diagnosis = src_dateevent) %>%
    select(-src_ctv3code, -src_ctv3text) %>%
    group_by(person_id) %>%
    arrange(date_stroke_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Valve Disease
valve_disease <- valve_disease %>%
    mutate(has_valve_disease = 1) %>%
    rename(date_valve_disease_diagnosis = src_dateevent) %>%
    select(-src_ctv3code, -src_ctv3text) %>%
    group_by(person_id) %>%
    arrange(date_valve_disease_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()





print("Angina")
head(angina)

print("Cardiomyopathy")
head(cardiomyopathy)

print("Heart Failure")
head(heart_failure)

print("Hypertension")
head(hypertension)

print("IHD")
head(ihd)

print("MI")
head(mi)

print("PVD")
head(pvd)

print("Stroke")
head(stroke)

print("Valve Disease")
head(valve_disease)

nrow(angina)
nrow(cardiomyopathy)
nrow(heart_failure)
nrow(hypertension)
nrow(ihd)
nrow(mi)
nrow(pvd)
nrow(stroke)
nrow(valve_disease)

# join all tables
cvd <- full_join(angina, cardiomyopathy, by = "person_id")
cvd <- full_join(cvd, heart_failure, by = "person_id")
cvd <- full_join(cvd, hypertension, by = "person_id")
cvd <- full_join(cvd, ihd, by = "person_id")
cvd <- full_join(cvd, mi, by = "person_id")
cvd <- full_join(cvd, pvd, by = "person_id")
cvd <- full_join(cvd, stroke, by = "person_id")
cvd <- full_join(cvd, valve_disease, by = "person_id")

# replace NA with 0 in has_* columns
cvd <- cvd %>% 
    mutate_at(c('has_angina','has_cardiomyopathy','has_heart_failure','has_hypertension','has_ihd','has_mi','has_pvd','has_stroke','has_valve_disease'), ~replace_na(.,0))

# sum number of cvds, add indicator for cvd
cvd <- cvd %>%
    mutate(has_cvd = ifelse(has_angina == 1 | has_cardiomyopathy == 1 | has_heart_failure == 1 | has_hypertension == 1 | has_ihd == 1 | has_mi == 1 | has_pvd == 1 | has_stroke == 1 | has_valve_disease == 1,1,0)) %>%
    mutate(cvd_number = rowSums(select(.,"has_angina", "has_cardiomyopathy", "has_heart_failure", "has_hypertension", "has_ihd", "has_mi", "has_pvd", "has_stroke", "has_valve_disease")))
head(cvd)
table(cvd$cvd_number)

# change date variable field class to date to be able to save dataset
cvd$date_angina_diagnosis <- as.Date(cvd$date_angina_diagnosis)
cvd$date_cardiomyopathy_diagnosis <- as.Date(cvd$date_cardiomyopathy_diagnosis)
cvd$date_valve_disease_diagnosis <- as.Date(cvd$date_valve_disease_diagnosis)
cvd$date_heart_failure_diagnosis <- as.Date(cvd$date_heart_failure_diagnosis)
cvd$date_hypertension_diagnosis <- as.Date(cvd$date_hypertension_diagnosis)
cvd$date_ihd_diagnosis <- as.Date(cvd$date_ihd_diagnosis)
cvd$date_mi_diagnosis <- as.Date(cvd$date_mi_diagnosis)
cvd$date_pvd_diagnosis <- as.Date(cvd$date_pvd_diagnosis)
cvd$date_stroke_diagnosis <- as.Date(cvd$date_stroke_diagnosis)

# add date of birth, gender, ethnicity and imd decile 
# person table
sql <- "SELECT * FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Patient_Denom`"
tb <- bq_project_query(projectid, sql)
person <-bq_table_download(tb)
head(person)

nrow(cvd)

# join person table to cvd table
cvd_pc <- person %>%
    right_join(cvd)

head(cvd_pc)

# Derive age at date of diagnosis for each cvd
cvd_pc <- cvd_pc %>%
    mutate(age_angina_diag = interval(birth_datetime, date_angina_diagnosis)/years(1)) %>%
    mutate(age_angina_diag = floor(age_angina_diag)) %>%
    mutate(age_cardiomyopathy_diag = interval(birth_datetime, date_cardiomyopathy_diagnosis)/years(1)) %>%
    mutate(age_cardiomyopathy_diag = floor(age_cardiomyopathy_diag)) %>%
    mutate(age_heart_failure_diag = interval(birth_datetime, date_heart_failure_diagnosis)/years(1)) %>%
    mutate(age_heart_failure_diag = floor(age_heart_failure_diag)) %>%
    mutate(age_hypertension_diag = interval(birth_datetime, date_hypertension_diagnosis)/years(1)) %>%
    mutate(age_hypertension_diag = floor(age_hypertension_diag)) %>%
    mutate(age_ihd_diag = interval(birth_datetime, date_ihd_diagnosis)/years(1)) %>%
    mutate(age_ihd_diag = floor(age_ihd_diag)) %>%
    mutate(age_mi_diag = interval(birth_datetime, date_mi_diagnosis)/years(1)) %>%
    mutate(age_mi_diag = floor(age_mi_diag)) %>%
    mutate(age_pvd_diag = interval(birth_datetime, date_pvd_diagnosis)/years(1)) %>%
    mutate(age_pvd_diag = floor(age_pvd_diag)) %>%
    mutate(age_stroke_diag = interval(birth_datetime, date_stroke_diagnosis)/years(1)) %>%
    mutate(age_stroke_diag = floor(age_stroke_diag)) %>%
    mutate(age_valve_disease_diag = interval(birth_datetime, date_valve_disease_diagnosis)/years(1)) %>%
    mutate(age_valve_disease_diag = floor(age_valve_disease_diag))


head(cvd_pc)

# remove birth date, gender, ethnicity and imd
cvd_pc <- cvd_pc %>%
    select(-birth_datetime, -gender, -ethnicity, -imd_decile)

colnames(cvd_pc)

cvd_pc <- cvd_pc %>%
mutate(in_data = 1)

nrow(cvd_pc)

table(cvd_pc$has_valve_disease)

# save
# set destination table - do this regardless of whether it already exists
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
dest_table <- bq_table(dest_dataset, "Cardiovascular_Diseases_Primary_Care")

# create a full table spec as follows
# then pass this to bq_table_upload
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("date_angina_diagnosis", "DATE"),
                              bq_field("has_angina", "INT64"),
                              bq_field("age_angina_diag", "INT64"),
                              bq_field("date_cardiomyopathy_diagnosis", "DATE"),
                              bq_field("has_cardiomyopathy", "INT64"),
                              bq_field("age_cardiomyopathy_diag", "INT64"),
                              bq_field("date_heart_failure_diagnosis", "DATE"),
                              bq_field("has_heart_failure", "INT64"),
                              bq_field("age_heart_failure_diag", "INT64"),
                              bq_field("date_hypertension_diagnosis", "DATE"),
                              bq_field("has_hypertension", "INT64"),
                              bq_field("age_hypertension_diag", "INT64"),
                              bq_field("date_ihd_diagnosis", "DATE"),
                              bq_field("has_ihd", "INT64"),
                              bq_field("age_ihd_diag", "INT64"),
                              bq_field("date_mi_diagnosis", "DATE"),
                              bq_field("has_mi", "INT64"),
                              bq_field("age_mi_diag", "INT64"),
                              bq_field("date_pvd_diagnosis", "DATE"),
                              bq_field("has_pvd", "INT64"),
                              bq_field("age_pvd_diag", "INT64"),
                              bq_field("date_stroke_diagnosis", "DATE"),
                              bq_field("has_stroke", "INT64"),
                              bq_field("age_stroke_diag", "INT64"),
                              bq_field("date_valve_disease_diagnosis", "DATE"),
                              bq_field("has_valve_disease", "INT64"),
                              bq_field("age_valve_disease_diag", "INT64"),
                              bq_field("has_cvd", "INT64"),
                              bq_field("cvd_number", "INT64"),
                             bq_field("in_data", "INT64")))

# delete an existing table if necessary
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
# upload file_data to dest_table
bq_table_upload(dest_table, cvd, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')
