==> code/6.CVD_SecondaryCare_Episodes.ipynb <==
library(bigrquery)
library(tidyverse)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# load episodes
# Angina: Secondary care between 2002-2022
sql1 <- "SELECT person_id, gender, birth_month, birth_year, episode_start_date, episode_end_date, diagnosis_1  FROM `yhcr-prd-phm-bia-core.CB_FDM_Warehouse_V2.tbl_episode` 
WHERE episode_start_date BETWEEN '2002-01-01' AND '2022-12-31' AND (diagnosis_1 LIKE '%I200%'
OR diagnosis_1 LIKE '%I248%'
OR diagnosis_1 LIKE '%I209%'
OR diagnosis_1 LIKE '%I20%'
OR diagnosis_1 LIKE '%A360%'
OR diagnosis_1 LIKE '%A691%'
OR diagnosis_1 LIKE '%B085%'
OR diagnosis_1 LIKE '%B27%'
OR diagnosis_1 LIKE '%D70X%'
OR diagnosis_1 LIKE '%I739%'
OR diagnosis_1 LIKE '%J020%'
OR diagnosis_1 LIKE '%J312%'
OR diagnosis_1 LIKE '%K122%'
OR diagnosis_1 LIKE '%I208%'
OR diagnosis_1 LIKE '%I201%')"
tb1 <- bq_project_query(projectid, sql1)
angina <-bq_table_download(tb1) # store data in tibble


# Cardiomyopathy: Secondary care between 2002-2022
sql2 <- "SELECT person_id, gender, birth_month, birth_year, episode_start_date, episode_end_date, diagnosis_1  FROM `yhcr-prd-phm-bia-core.CB_FDM_Warehouse_V2.tbl_episode` 
WHERE episode_start_date BETWEEN '2002-01-01' AND '2022-12-31' AND (diagnosis_1 LIKE '%E854D%'
OR diagnosis_1 LIKE '%I431A%'
OR diagnosis_1 LIKE '%I255%'
OR diagnosis_1 LIKE '%I429%'
OR diagnosis_1 LIKE '%I42%'
OR diagnosis_1 LIKE '%I090%'
OR diagnosis_1 LIKE '%I11%'
OR diagnosis_1 LIKE '%I251%'
OR diagnosis_1 LIKE '%I43A%'
OR diagnosis_1 LIKE '%O903%'
OR diagnosis_1 LIKE '%O994%'
OR diagnosis_1 LIKE '%Q248%'
OR diagnosis_1 LIKE '%I423%'
OR diagnosis_1 LIKE '%I421%'
OR diagnosis_1 LIKE '%I428%'
OR diagnosis_1 LIKE '%I425%'
OR diagnosis_1 LIKE '%I422%'
OR diagnosis_1 LIKE '%I420%'
OR diagnosis_1 LIKE '%B572D%'
OR diagnosis_1 LIKE '%B570D%'
OR diagnosis_1 LIKE '%I981A%'
OR diagnosis_1 LIKE '%I412A%'
OR diagnosis_1 LIKE '%I428D%'
OR diagnosis_1 LIKE '%E90XA%'
OR diagnosis_1 LIKE '%E740D%'
OR diagnosis_1 LIKE '%E763D%'
OR diagnosis_1 LIKE '%I528A%'
OR diagnosis_1 LIKE '%I438A%'
OR diagnosis_1 LIKE '%G111D%'
OR diagnosis_1 LIKE '%G711D%'
OR diagnosis_1 LIKE '%I427%'
OR diagnosis_1 LIKE '%I515%'
OR diagnosis_1 LIKE '%I430A%'
OR diagnosis_1 LIKE '%I432A%'
OR diagnosis_1 LIKE '%G710%'
OR diagnosis_1 LIKE '%I518%'
OR diagnosis_1 LIKE '%G710D%')"
tb2 <- bq_project_query(projectid, sql2)
cardiomyopathy <-bq_table_download(tb2) # store data in tibble

# Heart Failure: Secondary care between 2002-2022
sql3 <- "SELECT person_id, gender, birth_month, birth_year, episode_start_date, episode_end_date, diagnosis_1  FROM `yhcr-prd-phm-bia-core.CB_FDM_Warehouse_V2.tbl_episode` 
WHERE episode_start_date BETWEEN '2002-01-01' AND '2022-12-31' AND (diagnosis_1 LIKE '%I099%'
OR diagnosis_1 LIKE '%I509%'
OR diagnosis_1 LIKE '%I50%'
OR diagnosis_1 LIKE '%I110%'
OR diagnosis_1 LIKE '%I130%'
OR diagnosis_1 LIKE '%I132%'
OR diagnosis_1 LIKE '%I971%'
OR diagnosis_1 LIKE '%I978%'
OR diagnosis_1 LIKE '%O291%'
OR diagnosis_1 LIKE '%O742%'
OR diagnosis_1 LIKE '%O754%'
OR diagnosis_1 LIKE '%O891%'
OR diagnosis_1 LIKE '%P290%'
OR diagnosis_1 LIKE '%T818%'
OR diagnosis_1 LIKE '%E059D%'
OR diagnosis_1 LIKE '%I500%'
OR diagnosis_1 LIKE '%I10X%'
OR diagnosis_1 LIKE '%I501%'
OR diagnosis_1 LIKE '%I018%'
OR diagnosis_1 LIKE '%I020%'
OR diagnosis_1 LIKE '%I098%'
OR diagnosis_1 LIKE '%I970%')"
tb3 <- bq_project_query(projectid, sql3)
heart_failure <-bq_table_download(tb3) # store data in tibble

# Hypertension: Secondary care between 2002-2022
sql4 <- "SELECT person_id, gender, birth_month, birth_year, episode_start_date, episode_end_date, diagnosis_1  FROM `yhcr-prd-phm-bia-core.CB_FDM_Warehouse_V2.tbl_episode` 
WHERE episode_start_date BETWEEN '2002-01-01' AND '2022-12-31' AND (diagnosis_1 LIKE '%I158%'
OR diagnosis_1 LIKE '%O16X%'
OR diagnosis_1 LIKE '%I21%'
OR diagnosis_1 LIKE '%I22%'
OR diagnosis_1 LIKE '%I23%'
OR diagnosis_1 LIKE '%I24%'
OR diagnosis_1 LIKE '%I25%'
OR diagnosis_1 LIKE '%I270%'
OR diagnosis_1 LIKE '%O10%'
OR diagnosis_1 LIKE '%O11X%'
OR diagnosis_1 LIKE '%O13X%'
OR diagnosis_1 LIKE '%O14%'
OR diagnosis_1 LIKE '%O15%'
OR diagnosis_1 LIKE '%P292%'
OR diagnosis_1 LIKE '%I12%'
OR diagnosis_1 LIKE '%I13%'
OR diagnosis_1 LIKE '%P000%'
OR diagnosis_1 LIKE '%I129%'
OR diagnosis_1 LIKE '%I159%'
OR diagnosis_1 LIKE '%I15%'
OR diagnosis_1 LIKE '%O104%'
OR diagnosis_1 LIKE '%I150%'
OR diagnosis_1 LIKE '%I152%'
OR diagnosis_1 LIKE '%I151%'
OR diagnosis_1 LIKE '%O12%'
OR diagnosis_1 LIKE '%O100%'
OR diagnosis_1 LIKE '%O102%'
OR diagnosis_1 LIKE '%O109%'
OR diagnosis_1 LIKE '%O121%'
OR diagnosis_1 LIKE '%O141%'
OR diagnosis_1 LIKE '%O159%'
OR diagnosis_1 LIKE '%O152%'
OR diagnosis_1 LIKE '%O150%'
OR diagnosis_1 LIKE '%O149%'
OR diagnosis_1 LIKE '%R688%'
OR diagnosis_1 LIKE '%O151%'
OR diagnosis_1 LIKE '%O142%'
OR diagnosis_1 LIKE '%O140%')"
tb4 <- bq_project_query(projectid, sql4)
hypertension <-bq_table_download(tb4) # store data in tibble

# Ischaemic Heart Disease: Secondary care between 2002-2022
sql5 <- "SELECT person_id, gender, birth_month, birth_year, episode_start_date, episode_end_date, diagnosis_1  FROM `yhcr-prd-phm-bia-core.CB_FDM_Warehouse_V2.tbl_episode` 
WHERE episode_start_date BETWEEN '2002-01-01' AND '2022-12-31' AND (diagnosis_1 LIKE '%I219%'
OR diagnosis_1 LIKE '%I214%'
OR diagnosis_1 LIKE '%I212%'
OR diagnosis_1 LIKE '%I249%'
OR diagnosis_1 LIKE '%I258%'
OR diagnosis_1 LIKE '%I250%'
OR diagnosis_1 LIKE '%I253%'
OR diagnosis_1 LIKE '%I254%'
OR diagnosis_1 LIKE '%I259%'
OR diagnosis_1 LIKE '%I233%'
OR diagnosis_1 LIKE '%I256%'
OR diagnosis_1 LIKE '%P294%'
OR diagnosis_1 LIKE '%I232%')"
tb5 <- bq_project_query(projectid, sql5)
ihd <-bq_table_download(tb5) # store data in tibble

# Myocardial Infarction: Secondary care between 2002-2022
sql6 <- "SELECT person_id, gender, birth_month, birth_year, episode_start_date, episode_end_date, diagnosis_1  FROM `yhcr-prd-phm-bia-core.CB_FDM_Warehouse_V2.tbl_episode` 
WHERE episode_start_date BETWEEN '2002-01-01' AND '2022-12-31' AND (diagnosis_1 LIKE '%I210%'
OR diagnosis_1 LIKE '%I220%'
OR diagnosis_1 LIKE '%I211%'
OR diagnosis_1 LIKE '%I221%'
OR diagnosis_1 LIKE '%I229%'
OR diagnosis_1 LIKE '%I228%'
OR diagnosis_1 LIKE '%I213%'
OR diagnosis_1 LIKE '%I252%'
OR diagnosis_1 LIKE '%A520D%'
OR diagnosis_1 LIKE '%I520A%')"
tb6 <- bq_project_query(projectid, sql6)
mi <-bq_table_download(tb6) # store data in tibble

# Peripheral Vascular Disease: Secondary care between 2002-2022
sql7 <- "SELECT person_id, gender, birth_month, birth_year, episode_start_date, episode_end_date, diagnosis_1  FROM `yhcr-prd-phm-bia-core.CB_FDM_Warehouse_V2.tbl_episode` 
WHERE episode_start_date BETWEEN '2002-01-01' AND '2022-12-31' AND (diagnosis_1 LIKE '%I738%'
OR diagnosis_1 LIKE '%E145D%'
OR diagnosis_1 LIKE '%E105D%'
OR diagnosis_1 LIKE '%E115D%'
OR diagnosis_1 LIKE '%E125D%'
OR diagnosis_1 LIKE '%E135D%'
OR diagnosis_1 LIKE '%O243D%'
OR diagnosis_1 LIKE '%O244D%'
OR diagnosis_1 LIKE '%O249D%'
OR diagnosis_1 LIKE '%I792A%'
OR diagnosis_1 LIKE '%I73%')"
tb7 <- bq_project_query(projectid, sql7)
pvd <-bq_table_download(tb7) # store data in tibble

# Stroke: Secondary care between 2002-2022
sql8 <- "SELECT person_id, gender, birth_month, birth_year, episode_start_date, episode_end_date, diagnosis_1  FROM `yhcr-prd-phm-bia-core.CB_FDM_Warehouse_V2.tbl_episode` 
WHERE episode_start_date BETWEEN '2002-01-01' AND '2022-12-31' AND (diagnosis_1 LIKE '%I619%'
OR diagnosis_1 LIKE '%I61%'
OR diagnosis_1 LIKE '%I691%'
OR diagnosis_1 LIKE '%I618%'
OR diagnosis_1 LIKE '%G459%'
OR diagnosis_1 LIKE '%G45%'
OR diagnosis_1 LIKE '%G451%'
OR diagnosis_1 LIKE '%G450%'
OR diagnosis_1 LIKE '%R55X%'
OR diagnosis_1 LIKE '%I603%'
OR diagnosis_1 LIKE '%I611%'
OR diagnosis_1 LIKE '%I663D%'
OR diagnosis_1 LIKE '%G463A%'
OR diagnosis_1 LIKE '%I662D%'
OR diagnosis_1 LIKE '%G462A%'
OR diagnosis_1 LIKE '%I614%'
OR diagnosis_1 LIKE '%P526%'
OR diagnosis_1 LIKE '%S068%'
OR diagnosis_1 LIKE '%I631%'
OR diagnosis_1 LIKE '%I600%'
OR diagnosis_1 LIKE '%I639%'
OR diagnosis_1 LIKE '%Z866%'
OR diagnosis_1 LIKE '%I612%'
OR diagnosis_1 LIKE '%I64X%'
OR diagnosis_1 LIKE '%I610%'
OR diagnosis_1 LIKE '%I690%'
OR diagnosis_1 LIKE '%I692%'
OR diagnosis_1 LIKE '%I693%'
OR diagnosis_1 LIKE '%I694%'
OR diagnosis_1 LIKE '%I63%'
OR diagnosis_1 LIKE '%I669%'
OR diagnosis_1 LIKE '%I66%'
OR diagnosis_1 LIKE '%I633%'
OR diagnosis_1 LIKE '%I634%'
OR diagnosis_1 LIKE '%I635%'
OR diagnosis_1 LIKE '%G465A%'
OR diagnosis_1 LIKE '%I639D%'
OR diagnosis_1 LIKE '%I609%'
OR diagnosis_1 LIKE '%I60%'
OR diagnosis_1 LIKE '%P103%'
OR diagnosis_1 LIKE '%P525%'
OR diagnosis_1 LIKE '%I668%'
OR diagnosis_1 LIKE '%G452%'
OR diagnosis_1 LIKE '%I601%'
OR diagnosis_1 LIKE '%I660D%'
OR diagnosis_1 LIKE '%G460A%'
OR diagnosis_1 LIKE '%I613%'
OR diagnosis_1 LIKE '%I630%'
OR diagnosis_1 LIKE '%I632%'
OR diagnosis_1 LIKE '%G466A%'
OR diagnosis_1 LIKE '%G453%'
OR diagnosis_1 LIKE '%I616%'
OR diagnosis_1 LIKE '%G458%'
OR diagnosis_1 LIKE '%I606%'
OR diagnosis_1 LIKE '%I607%'
OR diagnosis_1 LIKE '%Z867%'
OR diagnosis_1 LIKE '%I615%'
OR diagnosis_1 LIKE '%G454%'
OR diagnosis_1 LIKE '%R413%'
OR diagnosis_1 LIKE '%I678%'
OR diagnosis_1 LIKE '%I636%'
OR diagnosis_1 LIKE '%I604%'
OR diagnosis_1 LIKE '%I661D%'
OR diagnosis_1 LIKE '%G461A%'
OR diagnosis_1 LIKE '%G467A%'
OR diagnosis_1 LIKE '%I638%'
OR diagnosis_1 LIKE '%I608%'
OR diagnosis_1 LIKE '%I605%'
OR diagnosis_1 LIKE '%P101%'
OR diagnosis_1 LIKE '%P524%'
OR diagnosis_1 LIKE '%G464A%'
OR diagnosis_1 LIKE '%I602%'
OR diagnosis_1 LIKE '%I679%'
OR diagnosis_1 LIKE '%I67%'
OR diagnosis_1 LIKE '%G46%'
OR diagnosis_1 LIKE '%I68%'
OR diagnosis_1 LIKE '%I671%'
OR diagnosis_1 LIKE '%I672%'
OR diagnosis_1 LIKE '%S062%'
OR diagnosis_1 LIKE '%S063%'
OR diagnosis_1 LIKE '%P10%'
OR diagnosis_1 LIKE '%P52%'
OR diagnosis_1 LIKE '%I658%')"
tb8 <- bq_project_query(projectid, sql8)
stroke <-bq_table_download(tb8) # store data in tibble

# Valve Disease: Secondary care between 2002-2022
sql9 <- "SELECT person_id, gender, birth_month, birth_year, episode_start_date, episode_end_date, diagnosis_1  FROM `yhcr-prd-phm-bia-core.CB_FDM_Warehouse_V2.tbl_episode` 
WHERE episode_start_date BETWEEN '2002-01-01' AND '2022-12-31' AND (diagnosis_1 LIKE '%I059%'
OR diagnosis_1 LIKE '%I05%'
OR diagnosis_1 LIKE '%I080%'
OR diagnosis_1 LIKE '%I081%'
OR diagnosis_1 LIKE '%I083%'
OR diagnosis_1 LIKE '%I34%'
OR diagnosis_1 LIKE '%I050%'
OR diagnosis_1 LIKE '%I051%'
OR diagnosis_1 LIKE '%I052%'
OR diagnosis_1 LIKE '%I011%'
OR diagnosis_1 LIKE '%I342%'
OR diagnosis_1 LIKE '%I069%'
OR diagnosis_1 LIKE '%I06%'
OR diagnosis_1 LIKE '%I060%'
OR diagnosis_1 LIKE '%I062%'
OR diagnosis_1 LIKE '%I082%'
OR diagnosis_1 LIKE '%I061%'
OR diagnosis_1 LIKE '%I078%'
OR diagnosis_1 LIKE '%I070%'
OR diagnosis_1 LIKE '%I072%'
OR diagnosis_1 LIKE '%I071%'
OR diagnosis_1 LIKE '%I079%'
OR diagnosis_1 LIKE '%I088%'
OR diagnosis_1 LIKE '%I234%'
OR diagnosis_1 LIKE '%I235%'
OR diagnosis_1 LIKE '%I340%'
OR diagnosis_1 LIKE '%Q233%'
OR diagnosis_1 LIKE '%I349%'
OR diagnosis_1 LIKE '%I359%'
OR diagnosis_1 LIKE '%I35%'
OR diagnosis_1 LIKE '%I391A%'
OR diagnosis_1 LIKE '%I351%'
OR diagnosis_1 LIKE '%Q231%'
OR diagnosis_1 LIKE '%I350%'
OR diagnosis_1 LIKE '%I352%'
OR diagnosis_1 LIKE '%Q230%'
OR diagnosis_1 LIKE '%I369%'
OR diagnosis_1 LIKE '%I36%'
OR diagnosis_1 LIKE '%I361%'
OR diagnosis_1 LIKE '%I360%'
OR diagnosis_1 LIKE '%I362%'
OR diagnosis_1 LIKE '%I07%'
OR diagnosis_1 LIKE '%I379%'
OR diagnosis_1 LIKE '%I37%'
OR diagnosis_1 LIKE '%I393A%'
OR diagnosis_1 LIKE '%Q221%'
OR diagnosis_1 LIKE '%Q222%'
OR diagnosis_1 LIKE '%Q223%'
OR diagnosis_1 LIKE '%I371%'
OR diagnosis_1 LIKE '%I370%'
OR diagnosis_1 LIKE '%I372%'
OR diagnosis_1 LIKE '%I089%'
OR diagnosis_1 LIKE '%I08%'
OR diagnosis_1 LIKE '%I394A%'
OR diagnosis_1 LIKE '%I38X%'
OR diagnosis_1 LIKE '%I398A%'
OR diagnosis_1 LIKE '%I512%'
OR diagnosis_1 LIKE '%I058%'
OR diagnosis_1 LIKE '%I068%'
OR diagnosis_1 LIKE '%I348%'
OR diagnosis_1 LIKE '%I358%'
OR diagnosis_1 LIKE '%I368%'
OR diagnosis_1 LIKE '%I378%'
OR diagnosis_1 LIKE '%I390A%'
OR diagnosis_1 LIKE '%I392A%'
OR diagnosis_1 LIKE '%Q220%'
OR diagnosis_1 LIKE '%Q224%'
OR diagnosis_1 LIKE '%Q225%'
OR diagnosis_1 LIKE '%Q232%'
OR diagnosis_1 LIKE '%Q238%'
OR diagnosis_1 LIKE '%Q244%'
OR diagnosis_1 LIKE '%Q253%'
OR diagnosis_1 LIKE '%Q228%'
OR diagnosis_1 LIKE '%Q229%'
OR diagnosis_1 LIKE '%T828%'
OR diagnosis_1 LIKE '%I341%'
OR diagnosis_1 LIKE '%Q239%'
OR diagnosis_1 LIKE '%Q23%'
OR diagnosis_1 LIKE '%I710%'
OR diagnosis_1 LIKE '%Q234%'
OR diagnosis_1 LIKE '%Q226%'
OR diagnosis_1 LIKE '%T820%'
OR diagnosis_1 LIKE '%I712%'
OR diagnosis_1 LIKE '%I711%'
OR diagnosis_1 LIKE '%Q254%'
OR diagnosis_1 LIKE '%I790A%'
OR diagnosis_1 LIKE '%Q208%'
OR diagnosis_1 LIKE '%Q210%'
OR diagnosis_1 LIKE '%I02%'
OR diagnosis_1 LIKE '%I511%'
OR diagnosis_1 LIKE '%I33%'
OR diagnosis_1 LIKE '%A520A%'
OR diagnosis_1 LIKE '%I330%')"
tb9 <- bq_project_query(projectid, sql9)
valve_disease <-bq_table_download(tb9) # store data in tibble

nrow(angina)
nrow(cardiomyopathy)
nrow(heart_failure)
nrow(hypertension)
nrow(ihd)
nrow(mi)
nrow(pvd)
nrow(stroke)
nrow(valve_disease)

head(angina)

# Keep earliest date for each
# Angina
angina <- angina %>%
    mutate(has_angina = 1) %>%
    rename(date_angina_diagnosis = episode_start_date) %>%
    select(-diagnosis_1) %>%
    group_by(person_id) %>%
    arrange(date_angina_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Cardiomyopathy
cardiomyopathy <- cardiomyopathy %>%
    mutate(has_cardiomyopathy = 1) %>%
    rename(date_cardiomyopathy_diagnosis = episode_start_date) %>%
    select(-diagnosis_1) %>%
    group_by(person_id) %>%
    arrange(date_cardiomyopathy_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Heart Failure
heart_failure <- heart_failure %>%
    mutate(has_heart_failure = 1) %>%
    rename(date_heart_failure_diagnosis = episode_start_date) %>%
    select(-diagnosis_1) %>%
    group_by(person_id) %>%
    arrange(date_heart_failure_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Hypertension
hypertension <- hypertension %>%
    mutate(has_hypertension = 1) %>%
    rename(date_hypertension_diagnosis = episode_start_date) %>%
    select(-diagnosis_1) %>%
    group_by(person_id) %>%
    arrange(date_hypertension_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Ischaemic Heart Disease
ihd <- ihd %>%
    mutate(has_ihd = 1) %>%
    rename(date_ihd_diagnosis = episode_start_date) %>%
    select(-diagnosis_1) %>%
    group_by(person_id) %>%
    arrange(date_ihd_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Myocardial Infarction
mi <- mi %>%
    mutate(has_mi = 1) %>%
    rename(date_mi_diagnosis = episode_start_date) %>%
    select(-diagnosis_1) %>%
    group_by(person_id) %>%
    arrange(date_mi_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Peripheral Vascular Disease
pvd <- pvd %>%
    mutate(has_pvd = 1) %>%
    rename(date_pvd_diagnosis = episode_start_date) %>%
    select(-diagnosis_1) %>%
    group_by(person_id) %>%
    arrange(date_pvd_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Stroke
stroke <- stroke %>%
    mutate(has_stroke = 1) %>%
    rename(date_stroke_diagnosis = episode_start_date) %>%
    select(-diagnosis_1) %>%
    group_by(person_id) %>%
    arrange(date_stroke_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()

# Valve Disease
valve_disease <- valve_disease %>%
    mutate(has_valve_disease = 1) %>%
    rename(date_valve_disease_diagnosis = episode_start_date) %>%
    select(-diagnosis_1) %>%
    group_by(person_id) %>%
    arrange(date_valve_disease_diagnosis) %>%
    mutate(event_number = row_number()) %>%
    slice_min(event_number) %>%
    select(-event_number) %>%
    ungroup()


nrow(angina)
nrow(cardiomyopathy)
nrow(heart_failure)
nrow(hypertension)
nrow(ihd)
nrow(mi)
nrow(pvd)
nrow(stroke)
nrow(valve_disease)

head(angina)

# person table
sql <- "SELECT * FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Patient_Denom`"
tb <- bq_project_query(projectid, sql)
person <-bq_table_download(tb)
head(person)

# change class of gender to integer in cvd episodes
angina$gender <- as.integer(angina$gender)
cardiomyopathy$gender <- as.integer(cardiomyopathy$gender)
heart_failure$gender <- as.integer(heart_failure$gender)
hypertension$gender <- as.integer(hypertension$gender)
ihd$gender <- as.integer(ihd$gender)
mi$gender <- as.integer(mi$gender)
pvd$gender <- as.integer(pvd$gender)
stroke$gender <- as.integer(stroke$gender)
valve_disease$gender <- as.integer(valve_disease$gender)

# join person to each table 
angina <- right_join(person, angina, by = c("person_id", "gender"))
cardiomyopathy <- right_join(person, cardiomyopathy, by = c("person_id", "gender"))
heart_failure <- right_join(person, heart_failure, by = c("person_id", "gender"))
hypertension <- right_join(person, hypertension, by = c("person_id", "gender"))
ihd <- right_join(person, ihd, by = c("person_id", "gender"))
mi <- right_join(person, mi, by = c("person_id", "gender"))
pvd <- right_join(person, pvd, by = c("person_id", "gender"))
stroke <- right_join(person, stroke, by = c("person_id", "gender"))
valve_disease <- right_join(person, valve_disease, by = c("person_id", "gender"))

head(angina)

# calculate age at diagnosis and remove birth year and add length of stay 
angina_person <- angina %>%
    rename(date_angina_departure = episode_end_date) %>% 
    mutate(los_angina_days = round(difftime(date_angina_departure, date_angina_diagnosis, units="days"))) %>%
    mutate(age_angina_diag = interval(birth_datetime, date_angina_diagnosis)/years(1)) %>%
    mutate(age_angina_diag = floor(age_angina_diag)) %>%
    select(-birth_datetime, -birth_month, -birth_year)
angina_person$los_angina_days <- as.integer(angina_person$los_angina_days)
head(angina_person)

cardiomyopathy_person <- cardiomyopathy %>%
    rename(date_cardiomyopathy_departure = episode_end_date) %>% 
    mutate(los_cardiomyopathy_days = round(difftime(date_cardiomyopathy_departure, date_cardiomyopathy_diagnosis, units="days"))) %>%
    mutate(age_cardiomyopathy_diag = interval(birth_datetime, date_cardiomyopathy_diagnosis)/years(1)) %>%
    mutate(age_cardiomyopathy_diag = floor(age_cardiomyopathy_diag)) %>%
    select(-birth_datetime, -birth_month, -birth_year)
cardiomyopathy_person$los_cardiomyopathy_days <- as.integer(cardiomyopathy_person$los_cardiomyopathy_days)

heart_failure_person <- heart_failure %>%
  rename(date_heart_failure_departure = episode_end_date) %>% 
  mutate(los_heart_failure_days = round(difftime(date_heart_failure_departure, date_heart_failure_diagnosis, units="days"))) %>%
  mutate(age_heart_failure_diag = interval(birth_datetime, date_heart_failure_diagnosis)/years(1)) %>%
  mutate(age_heart_failure_diag = floor(age_heart_failure_diag)) %>%
  select(-birth_datetime, -birth_month, -birth_year)
heart_failure_person$los_heart_failure_days <- as.integer(heart_failure_person$los_heart_failure_days)

hypertension_person <- hypertension %>%
  rename(date_hypertension_departure = episode_end_date) %>% 
  mutate(los_hypertension_days = round(difftime(date_hypertension_departure, date_hypertension_diagnosis, units="days"))) %>%
  mutate(age_hypertension_diag = interval(birth_datetime, date_hypertension_diagnosis)/years(1)) %>%
  mutate(age_hypertension_diag = floor(age_hypertension_diag)) %>%
  select(-birth_datetime, -birth_month, -birth_year)
hypertension_person$los_hypertension_days <- as.integer(hypertension_person$los_hypertension_days)

ihd_person <- ihd %>%
  rename(date_ihd_departure = episode_end_date) %>% 
  mutate(los_ihd_days = round(difftime(date_ihd_departure, date_ihd_diagnosis, units="days"))) %>%
  mutate(age_ihd_diag = interval(birth_datetime, date_ihd_diagnosis)/years(1)) %>%
  mutate(age_ihd_diag = floor(age_ihd_diag)) %>%
  select(-birth_datetime, -birth_month, -birth_year)
ihd_person$los_ihd_days <- as.integer(ihd_person$los_ihd_days)

mi_person <- mi %>%
  rename(date_mi_departure = episode_end_date) %>% 
  mutate(los_mi_days = round(difftime(date_mi_departure, date_mi_diagnosis, units="days"))) %>%
  mutate(age_mi_diag = interval(birth_datetime, date_mi_diagnosis)/years(1)) %>%
  mutate(age_mi_diag = floor(age_mi_diag)) %>%
  select(-birth_datetime, -birth_month, -birth_year)
mi_person$los_mi_days <- as.integer(mi_person$los_mi_days)

pvd_person <- pvd %>%
  rename(date_pvd_departure = episode_end_date) %>% 
  mutate(los_pvd_days = round(difftime(date_pvd_departure, date_pvd_diagnosis, units="days"))) %>%
  mutate(age_pvd_diag = interval(birth_datetime, date_pvd_diagnosis)/years(1)) %>%
  mutate(age_pvd_diag = floor(age_pvd_diag)) %>%
  select(-birth_datetime, -birth_month, -birth_year)
pvd_person$los_pvd_days <- as.integer(pvd_person$los_pvd_days)

stroke_person <- stroke %>%
  rename(date_stroke_departure = episode_end_date) %>% 
  mutate(los_stroke_days = round(difftime(date_stroke_departure, date_stroke_diagnosis, units="days"))) %>%
  mutate(age_stroke_diag = interval(birth_datetime, date_stroke_diagnosis)/years(1)) %>%
  mutate(age_stroke_diag = floor(age_stroke_diag)) %>%
  select(-birth_datetime, -birth_month, -birth_year)
stroke_person$los_stroke_days <- as.integer(stroke_person$los_stroke_days)

valve_disease_person <- valve_disease %>%
  rename(date_valve_disease_departure = episode_end_date) %>% 
  mutate(los_valve_disease_days = round(difftime(date_valve_disease_departure, date_valve_disease_diagnosis, units="days"))) %>%
  mutate(age_valve_disease_diag = interval(birth_datetime, date_valve_disease_diagnosis)/years(1)) %>%
  mutate(age_valve_disease_diag = floor(age_valve_disease_diag)) %>%
  select(-birth_datetime, -birth_month, -birth_year)
valve_disease_person$los_valve_disease_days <- as.integer(valve_disease_person$los_valve_disease_days)

# join all cvd episodes tables together
cvd_ep <- full_join(angina_person, cardiomyopathy_person, by = c("person_id", "gender", "ethnicity", "imd_decile"))
cvd_ep <- full_join(cvd_ep, heart_failure_person, by = c("person_id", "gender", "ethnicity", "imd_decile"))
cvd_ep <- full_join(cvd_ep, hypertension_person, by = c("person_id", "gender", "ethnicity", "imd_decile"))
cvd_ep <- full_join(cvd_ep, ihd_person, by = c("person_id", "gender", "ethnicity", "imd_decile"))
cvd_ep <- full_join(cvd_ep, mi_person, by = c("person_id", "gender", "ethnicity", "imd_decile"))
cvd_ep <- full_join(cvd_ep, pvd_person, by = c("person_id", "gender", "ethnicity", "imd_decile"))
cvd_ep <- full_join(cvd_ep, stroke_person, by = c("person_id", "gender", "ethnicity", "imd_decile"))
cvd_ep <- full_join(cvd_ep, valve_disease_person, by = c("person_id", "gender", "ethnicity", "imd_decile"))
head(cvd_ep)

nrow(cvd_ep)

# has cvd and cvd number
cvd_ep <- cvd_ep %>%
    mutate_at(c('has_angina','has_cardiomyopathy','has_heart_failure','has_hypertension','has_ihd','has_mi','has_pvd','has_stroke','has_valve_disease'), ~replace_na(.,0)) %>%
    mutate(has_cvd = ifelse(has_angina == 1 | has_cardiomyopathy == 1 | has_heart_failure == 1 | has_hypertension == 1 | has_ihd == 1 | has_mi == 1 | has_pvd == 1 | has_stroke == 1 | has_valve_disease == 1,1,0)) %>%
    mutate(cvd_number = rowSums(select(.,"has_angina", "has_cardiomyopathy", "has_heart_failure", "has_hypertension", "has_ihd", "has_mi", "has_pvd", "has_stroke", "has_valve_disease")))
head(cvd_ep)

# create indicator for episodes
cvd_ep <- cvd_ep %>%
    mutate(in_data = 2)

head(cvd_ep)

cvd_ep <- cvd_ep %>%
    select(-gender, -ethnicity, -imd_decile)

colnames(cvd_ep)

# save dataset
# set destination table - do this regardless of whether it already exists
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
dest_table <- bq_table(dest_dataset, "Cardiovascular_Diseases_Episodes")

# create a full table spec as follows
# then pass this to bq_table_upload
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("age_angina_diag", "INT64"),
                              bq_field("date_angina_diagnosis", "DATE"),
                              bq_field("date_angina_departure", "DATE"),
                              bq_field("has_angina", "INT64"),
                              bq_field("los_angina_days", "INT64"),
                              bq_field("age_cardiomyopathy_diag", "INT64"),
                              bq_field("date_cardiomyopathy_diagnosis", "DATE"),
                              bq_field("date_cardiomyopathy_departure", "DATE"),
                              bq_field("has_cardiomyopathy", "INT64"),
                              bq_field("los_cardiomyopathy_days", "INT64"),
                              bq_field("age_heart_failure_diag", "INT64"),
                              bq_field("date_heart_failure_diagnosis", "DATE"),
                              bq_field("date_heart_failure_departure", "DATE"),
                              bq_field("has_heart_failure", "INT64"),
                              bq_field("los_heart_failure_days", "INT64"),
                              bq_field("age_hypertension_diag", "INT64"),
                              bq_field("date_hypertension_diagnosis", "DATE"),
                              bq_field("date_hypertension_departure", "DATE"),
                              bq_field("has_hypertension", "INT64"),
                              bq_field("los_hypertension_days", "INT64"),
                              bq_field("age_ihd_diag", "INT64"),
                              bq_field("date_ihd_diagnosis", "DATE"),
                              bq_field("date_ihd_departure", "DATE"),
                              bq_field("has_ihd", "INT64"),
                              bq_field("los_ihd_days", "INT64"),
                              bq_field("age_mi_diag", "INT64"),
                              bq_field("date_mi_diagnosis", "DATE"),
                              bq_field("date_mi_departure", "DATE"),
                              bq_field("has_mi", "INT64"),
                              bq_field("los_mi_days", "INT64"),
                              bq_field("age_pvd_diag", "INT64"),
                              bq_field("date_pvd_diagnosis", "DATE"),
                              bq_field("date_pvd_departure", "DATE"),
                              bq_field("has_pvd", "INT64"),
                              bq_field("los_pvd_days", "INT64"),
                              bq_field("age_stroke_diag", "INT64"),
                              bq_field("date_stroke_diagnosis", "DATE"),
                              bq_field("date_stroke_departure", "DATE"),
                              bq_field("has_stroke", "INT64"),
                              bq_field("los_stroke_days", "INT64"),
                              bq_field("age_valve_disease_diag", "INT64"),
                              bq_field("date_valve_disease_diagnosis", "DATE"),
                              bq_field("date_valve_disease_departure", "DATE"),
                              bq_field("has_valve_disease", "INT64"),
                              bq_field("los_valve_disease_days", "INT64"),
                              bq_field("has_cvd", "INT64"),
                              bq_field("cvd_number", "INT64"),
                                bq_field("in_data", "INT64")))

# delete an existing table if necessary
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
# upload file_data to dest_table
bq_table_upload(dest_table, cvd_ep, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')
