==> code/7.Joining_Covariates_Diabetes.ipynb <==
library(bigrquery)
library(tidyverse)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# load diabetes and covariates
# Diabetes
sql1 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Diabetes`"
tb1 <- bq_project_query(projectid, sql1)
diabetes <-bq_table_download(tb1)

# Cholesterol
sql2 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Cholesterol`"
tb2 <- bq_project_query(projectid, sql2)
cholesterol <-bq_table_download(tb2)

# BMI
sql3 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.BMI`"
tb3 <- bq_project_query(projectid, sql3)
bmi <-bq_table_download(tb3)

# Smoking
sql4 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Smoking`"
tb4 <- bq_project_query(projectid, sql4)
smoking <-bq_table_download(tb4)

# Blood Pressure
sql5 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Blood_Pressure`"
tb5 <- bq_project_query(projectid, sql5)
bp <-bq_table_download(tb5)

# Renal Disease: Creatinine
sql6 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Creatinine`"
tb6 <- bq_project_query(projectid, sql6)
creatinine <-bq_table_download(tb6)

# Renal Disease: ACR
sql7 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.ACR`"
tb7 <- bq_project_query(projectid, sql7)
acr <-bq_table_download(tb7)

# Renal Disease: eGFR
sql8 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.eGFR`"
tb8 <- bq_project_query(projectid, sql8)
egfr <-bq_table_download(tb8)

# drop BMI values <14 or >50
bmi <- bmi %>% filter(!bmi_value < 14) %>% filter(!bmi_value > 50)
nrow(bmi)

# Join diabetes and BMI tables
diab_bmi <- left_join(diabetes, bmi, by = "person_id")

# add indicator for bmi event closest to date of diabetes diagnosis, drop those that are > 365 days different
diab_bmi <- diab_bmi %>%
    mutate(diff_date = as.integer(abs(earliest_diabetes_diag - date_bmi))) %>%
    filter(!diff_date >= 366) %>%
    group_by(person_id) %>%
    arrange(person_id, diff_date) %>%
    mutate(order = row_number()) %>%
    ungroup() %>%
    mutate(closest_bmi = ifelse(order == 1, 1,0)) %>%
    filter(closest_bmi == 1)
head(diab_bmi)
sum(duplicated(diab_bmi))

# subset bmi 
BMI <- diab_bmi %>%
select(person_id, date_bmi, bmi, closest_bmi, bmi_value) %>%
filter(!is.na(date_bmi))
head(BMI)
sum(duplicated(BMI))

nrow(BMI)
table(BMI$bmi)

# join diabetes and chol
diab_chol <- left_join(diabetes, cholesterol, by = "person_id")
# add indicator for chol event closest to date of diabetes diagnosis, drop those that are > 365 days different
diab_chol <- diab_chol %>%
    mutate(diff_date = as.integer(abs(earliest_diabetes_diag - date_cholesterol))) %>%
    filter(!diff_date >= 366) %>%
    group_by(person_id) %>%
    arrange(person_id, diff_date) %>%
    mutate(order = row_number()) %>%
    ungroup() %>%
    mutate(closest_cholesterol = ifelse(order == 1, 1,0)) %>%
    filter(closest_cholesterol == 1)
head(diab_chol)

# subset chol 
Chol <- diab_chol %>%
select(person_id, date_cholesterol, cholesterol, closest_cholesterol) %>%
filter(!is.na(date_cholesterol))
head(Chol)
nrow(Chol)

table(Chol$cholesterol)

# join diabetes and smoking
diab_smoke <- left_join(diabetes, smoking, by = "person_id")
# add indicator for smoke event closest to date of diabetes diagnosis, drop those that are > 365 days different
diab_smoke <- diab_smoke %>%
    mutate(diff_date = as.integer(abs(earliest_diabetes_diag - date_smoking))) %>%
    filter(!diff_date >= 366) %>%
    group_by(person_id) %>%
    arrange(person_id, diff_date) %>%
    mutate(order = row_number()) %>%
    ungroup() %>%
    mutate(closest_smoking = ifelse(order == 1, 1,0)) %>%
    filter(closest_smoking == 1)
head(diab_smoke)

# subset smoking 
Smoke <- diab_smoke %>%
select(person_id, date_smoking, smoking, closest_smoking) %>%
filter(!is.na(date_smoking))
head(Smoke)
nrow(Smoke)

table(Smoke$smoking)

# join diabetes and BP
diab_bp <- left_join(diabetes, bp, by = "person_id")
# add indicator for bp event closest to date of diabetes diagnosis, drop those that are > 365 days different
diab_bp <- diab_bp %>%
    mutate(diff_date = as.integer(abs(earliest_diabetes_diag - date_bp))) %>%
    filter(!diff_date >= 366) %>%
    group_by(person_id) %>%
    arrange(person_id, diff_date) %>%
    mutate(order = row_number()) %>%
    ungroup() %>%
    mutate(closest_bp = ifelse(order == 1, 1,0)) %>%
    filter(closest_bp == 1)
head(diab_bp)
sum(duplicated(diab_bp))

# subset bp 
BP <- diab_bp %>%
select(person_id, date_bp, blood_pressure, sbp_value, dbp_value, closest_bp) %>%
filter(!is.na(date_bp))
head(BP)
nrow(BP)
sum(duplicated(BP))

table(BP$blood_pressure)

# join diabetes and renal
#diab_renal <- left_join(diabetes, renal, by = "person_id")
# keep date of bmi event closest to date of diabetes diagnosis
#diab_renal <- diab_renal %>%
#    mutate(diff_date = abs(date_earliest_diabetes - date_renal)) %>%
#    group_by(person_id) %>%
#    slice_min(diff_date, with_ties = FALSE)
#head(diab_renal)


# Join diabetes and creatinine
diab_cr <- left_join(diabetes, creatinine, by = "person_id")

# add indicator for creatinine event closest to date of diabetes diagnosis, drop those that are > 365 days different
diab_cr <- diab_cr %>%
    mutate(diff_date = as.integer(abs(earliest_diabetes_diag - date_creatinine))) %>%
    filter(!diff_date >= 366) %>%
    group_by(person_id) %>%
    arrange(person_id, diff_date) %>%
    mutate(order = row_number()) %>%
    ungroup() %>%
    mutate(closest_creatinine = ifelse(order == 1, 1,0)) %>%
    filter(closest_creatinine == 1)
head(diab_cr)

# subset cr 
Creatinine <- diab_cr %>%
select(person_id, date_creatinine, creatinine, closest_creatinine) %>%
filter(!is.na(date_creatinine))
head(Creatinine)
nrow(Creatinine)

table(Creatinine$creatinine)

# Join diabetes and albumin:creatinine ratio
diab_acr <- left_join(diabetes, acr, by = "person_id")

# add indicator for acr event closest to date of diabetes diagnosis, drop those that are > 365 days different
diab_acr <- diab_acr %>%
    mutate(diff_date = as.integer(abs(earliest_diabetes_diag - date_acr))) %>%
    filter(!diff_date >= 366) %>%
    group_by(person_id) %>%
    arrange(person_id, diff_date) %>%
    mutate(order = row_number()) %>%
    ungroup() %>%
    mutate(closest_acr = ifelse(order == 1, 1,0)) %>%
    filter(closest_acr == 1)
head(diab_acr)

# subset albumin:creatinine ratio
ACR <- diab_acr %>%
select(person_id, date_acr, acr, closest_acr) %>%
filter(!is.na(date_acr))
head(ACR)
nrow(ACR)

table(ACR$acr)

# Join diabetes and egfr
diab_egfr <- left_join(diabetes, egfr, by = "person_id")

# add indicator for egfr event closest to date of diabetes diagnosis, drop those that are > 365 days different
diab_egfr <- diab_egfr %>%
    mutate(diff_date = as.integer(abs(earliest_diabetes_diag - date_egfr))) %>%
    filter(!diff_date >= 366) %>%
    group_by(person_id) %>%
    arrange(person_id, diff_date) %>%
    mutate(order = row_number()) %>%
    ungroup() %>%
    mutate(closest_egfr = ifelse(order == 1, 1,0)) %>%
    filter(closest_egfr == 1)
head(diab_egfr)

# subset egfr 
EGFR <- diab_egfr %>%
select(person_id, date_egfr, egfr, closest_egfr) %>%
filter(!is.na(date_egfr))
head(EGFR)
nrow(EGFR)

table(EGFR$egfr)

# join all covariates to patient table
# import patient table
sql <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Patient_Denom`"
tb <- bq_project_query(projectid, sql)
patient_denom <-bq_table_download(tb)
head(patient_denom)

# join Chol, BMI, Smoke, BP and Renal_D to patient_denom
patient <- left_join(patient_denom, Chol)
patient <- left_join(patient, BMI)
patient <- left_join(patient, Smoke)
patient <- left_join(patient, BP)
patient <- left_join(patient, Creatinine)
patient <- left_join(patient, ACR)
patient <- left_join(patient, EGFR)

head(patient)
nrow(patient)
sum(!is.na(patient$cholesterol))
sum(!is.na(patient$bmi))
sum(!is.na(patient$smoking))
sum(!is.na(patient$blood_pressure))
sum(!is.na(patient$creatinine))
sum(!is.na(patient$acr))
sum(!is.na(patient$egfr))

patient_covs <- patient %>%
    mutate_at(c('cholesterol','bmi', 'smoking', 'blood_pressure', 'creatinine', 'acr', 'egfr'), ~replace_na(.,0))

table(patient_covs$gender)
table(patient_covs$ethnicity)
table(patient_covs$imd_decile)
table(patient_covs$cholesterol)
table(patient_covs$bmi)
table(patient_covs$smoking)
table(patient_covs$blood_pressure)
table(patient_covs$creatinine)
table(patient_covs$acr)
table(patient_covs$egfr)

rm(list=setdiff(ls(), "patient_covs"))
ls()

nrow(patient_covs)
table(patient_covs$t1d_status)

#colnames(patient_covs)
#patient_bp_bmi <- patient_covs %>%
#    select(person_id, birth_datetime, death_datetime, gender, ethnicity, imd_decile, date_bmi, bmi, bmi_value, closest_bmi, date_bp, blood_pressure, sbp_value, dbp_value, closest_bp)
#patient_covs <- patient_covs %>%
#    select(-date_bmi, -bmi, -bmi_value, -closest_bmi, -date_bp, -blood_pressure, -sbp_value, -dbp_value, -closest_bp)

# save patient_covs as csv
write.csv(patient_covs, '/home/jupyter/T2D_CVD_Study/Patient_Covariates_Table.csv')

sum(duplicated(patient_covs))

bq_auth()

colnames(patient_covs)

# save new patient table
# set destination table - do this regardless of whether it already exists
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
dest_table <- bq_table(dest_dataset, "Patient_Covariates_Table_1")

# create a full table spec as follows
# then pass this to bq_table_upload
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("birth_datetime", "DATE"),
                              bq_field("death_datetime", "DATE"),
                              bq_field("gender", "INT64"),
                              bq_field("ethnicity", "INT64"),
                              bq_field("imd_decile", "INT64"),
                              bq_field("date_cholesterol", "DATE"),
                              bq_field("cholesterol", "INT64"),
                              bq_field("closest_cholesterol", "INT64"),
                              bq_field("date_bmi", "DATE"),
                              bq_field("bmi", "INT64"),
                              bq_field("bmi_value", "FLOAT64"),
                              bq_field("closest_bmi", "INT64"),
                              bq_field("date_smoking", "DATE"),
                              bq_field("smoking", "INT64"),
                              bq_field("closest_smoking", "INT64"),
                              bq_field("date_bp", "DATE"),
                              bq_field("blood_pressure", "INT64"),
                              bq_field("sbp_value", "FLOAT64"),
                              bq_field("dbp_value", "FLOAT64"),
                              bq_field("closest_bp", "INT64"),
                              bq_field("date_creatinine", "DATE"),
                              bq_field("creatinine", "INT64"),
                              bq_field("closest_creatinine", "INT64"),
                              bq_field("date_acr", "DATE"),
                              bq_field("acr", "INT64"),
                              bq_field("closest_acr", "INT64"),
                              bq_field("date_egfr", "DATE"),
                              bq_field("egfr", "INT64"),
                              bq_field("closest_egfr", "INT64")))

# delete an existing table if necessary
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
# upload file_data to dest_table
bq_table_upload(dest_table, patient_covs, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')

# save new patient table
# set destination table - do this regardless of whether it already exists
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
dest_table <- bq_table(dest_dataset, "Patient_Covariates_Table1")

# create a full table spec as follows
# then pass this to bq_table_upload
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("birth_datetime", "DATE"),
                              bq_field("death_datetime", "DATE"),
                              bq_field("gender", "INT64"),
                              bq_field("ethnicity", "INT64"),
                              bq_field("imd_decile", "INT64"),
                              bq_field("date_cholesterol", "DATE"),
                              bq_field("cholesterol", "INT64"),
                              bq_field("closest_cholesterol", "INT64"),
                              bq_field("date_smoking", "DATE"),
                              bq_field("smoking", "INT64"),
                              bq_field("closest_smoking", "INT64"),
                              bq_field("date_creatinine", "DATE"),
                              bq_field("creatinine", "INT64"),
                              bq_field("closest_creatinine", "INT64"),
                              bq_field("date_acr", "DATE"),
                              bq_field("acr", "INT64"),
                              bq_field("closest_acr", "INT64"),
                              bq_field("date_egfr", "DATE"),
                              bq_field("egfr", "INT64"),
                              bq_field("closest_egfr", "INT64")))

# delete an existing table if necessary
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
# upload file_data to dest_table
bq_table_upload(dest_table, patient_covs, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')

# save new patient table
# set destination table - do this regardless of whether it already exists
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
dest_table <- bq_table(dest_dataset, "Patient_Covariates_Table_2")

# create a full table spec as follows
# then pass this to bq_table_upload
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("birth_datetime", "DATE"),
                              bq_field("death_datetime", "DATE"),
                              bq_field("gender", "INT64"),
                              bq_field("ethnicity", "INT64"),
                              bq_field("imd_decile", "INT64"),
                              bq_field("date_bmi", "DATE"),
                              bq_field("bmi", "INT64"),
                              bq_field("bmi_value", "FLOAT64"),
                              bq_field("closest_bmi", "INT64"),
                              bq_field("date_bp", "DATE"),
                              bq_field("blood_pressure", "INT64"),
                              bq_field("sbp_value", "FLOAT64"),
                              bq_field("dbp_value", "FLOAT64"),
                              bq_field("closest_bp", "INT64")))

# delete an existing table if necessary
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
# upload file_data to dest_table
bq_table_upload(dest_table, patient_bp_bmi, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')
