==> code/8.Join_CVD_Tables.ipynb <==
library(bigrquery)
library(tidyverse)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# load cvd tables
# CVD Primary Care
sql1 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Cardiovascular_Diseases_Primary_Care`"
tb1 <- bq_project_query(projectid, sql1)
cvd_pc <-bq_table_download(tb1)

# CVD A&E
sql2 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Cardiovascular_Diseases_AE`"
tb2 <- bq_project_query(projectid, sql2)
cvd_ae <-bq_table_download(tb2)

# CVD Episodes
sql3 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Cardiovascular_Diseases_Episodes`"
tb3 <- bq_project_query(projectid, sql3)
cvd_ep <-bq_table_download(tb3)


# Keep personid, date diagnosis and has_* variables
# Primary Care
pc_angina <- cvd_pc %>%
    select(person_id, date_angina_diagnosis, has_angina)
pc_cardiomyopathy <- cvd_pc %>%
    select(person_id, date_cardiomyopathy_diagnosis, has_cardiomyopathy)
pc_hf <- cvd_pc %>%
    select(person_id, date_heart_failure_diagnosis, has_heart_failure)
pc_ht <- cvd_pc %>%
    select(person_id, date_hypertension_diagnosis, has_hypertension)
pc_ihd <- cvd_pc %>%
    select(person_id, date_ihd_diagnosis, has_ihd)
pc_mi <- cvd_pc %>%
    select(person_id, date_mi_diagnosis, has_mi)
pc_pvd <- cvd_pc %>%
    select(person_id, date_pvd_diagnosis, has_pvd)
pc_stroke <- cvd_pc %>%
    select(person_id, date_stroke_diagnosis, has_stroke)
pc_vd <- cvd_pc %>%
    select(person_id, date_valve_disease_diagnosis, has_valve_disease)

# A&E
ae_angina <- cvd_ae %>%
    select(person_id, date_angina_diagnosis, has_angina)
ae_cardiomyopathy <- cvd_ae %>%
    select(person_id, date_cardiomyopathy_diagnosis, has_cardiomyopathy)
ae_hf <- cvd_ae %>%
    select(person_id, date_heart_failure_diagnosis, has_heart_failure)
ae_ht <- cvd_ae %>%
    select(person_id, date_hypertension_diagnosis, has_hypertension)
ae_ihd <- cvd_ae %>%
    select(person_id, date_ihd_diagnosis, has_ihd)
ae_mi <- cvd_ae %>%
    select(person_id, date_mi_diagnosis, has_mi)
ae_pvd <- cvd_ae %>%
    select(person_id, date_pvd_diagnosis, has_pvd)
ae_stroke <- cvd_ae %>%
    select(person_id, date_stroke_diagnosis, has_stroke)
ae_vd <- cvd_ae %>%
    select(person_id, date_valve_disease_diagnosis, has_valve_disease)

# Episodes
ep_angina <- cvd_ep %>%
    select(person_id, date_angina_diagnosis, has_angina)
ep_cardiomyopathy <- cvd_ep %>%
    select(person_id, date_cardiomyopathy_diagnosis, has_cardiomyopathy)
ep_hf <- cvd_ep %>%
    select(person_id, date_heart_failure_diagnosis, has_heart_failure)
ep_ht <- cvd_ep %>%
    select(person_id, date_hypertension_diagnosis, has_hypertension)
ep_ihd <- cvd_ep %>%
    select(person_id, date_ihd_diagnosis, has_ihd)
ep_mi <- cvd_ep %>%
    select(person_id, date_mi_diagnosis, has_mi)
ep_pvd <- cvd_ep %>%
    select(person_id, date_pvd_diagnosis, has_pvd)
ep_stroke <- cvd_ep %>%
    select(person_id, date_stroke_diagnosis, has_stroke)
ep_vd <- cvd_ep %>%
    select(person_id, date_valve_disease_diagnosis, has_valve_disease)

# join each cvd together
# angina
angina <- rbind(pc_angina, ae_angina, ep_angina)
cardiomyopathy <- rbind(pc_cardiomyopathy, ae_cardiomyopathy, ep_cardiomyopathy) 
heart_failure <- rbind(pc_hf, ae_hf, ep_hf)
hypertension <- rbind(pc_ht, ae_ht, ep_ht)
ihd <- rbind(pc_ihd, ae_ihd, ep_ihd)
mi <- rbind(pc_mi, ae_mi, ep_mi)
pvd <- rbind(pc_pvd, ae_pvd, ep_pvd)
stroke <- rbind(pc_stroke, ae_stroke, ep_stroke)
valve_disease <- rbind(pc_vd, ae_vd, ep_vd)

# Keep first occurrence for each cvd
# Angina
angina <- angina %>%
    filter(!has_angina == 0)
nrow(angina)
angina <- angina %>%
    group_by(person_id) %>%
    slice_min(date_angina_diagnosis, with_ties = FALSE)
nrow(angina)

# Cardiomyopathy
cardiomyopathy <- cardiomyopathy %>%
    filter(!has_cardiomyopathy == 0)
nrow(cardiomyopathy)
cardiomyopathy <- cardiomyopathy %>%
    group_by(person_id) %>%
    slice_min(date_cardiomyopathy_diagnosis, with_ties = FALSE)
nrow(cardiomyopathy)

# heart_failure
heart_failure <- heart_failure %>%
    filter(!has_heart_failure == 0)
nrow(heart_failure)
heart_failure <- heart_failure %>%
    group_by(person_id) %>%
    slice_min(date_heart_failure_diagnosis, with_ties = FALSE)
nrow(heart_failure)

# hypertension
hypertension <- hypertension %>%
    filter(!has_hypertension == 0)
nrow(hypertension)
hypertension <- hypertension %>%
    group_by(person_id) %>%
    slice_min(date_hypertension_diagnosis, with_ties = FALSE)
nrow(hypertension)

# ihd
ihd <- ihd %>%
    filter(!has_ihd == 0)
nrow(ihd)
ihd <- ihd %>%
    group_by(person_id) %>%
    slice_min(date_ihd_diagnosis, with_ties = FALSE)
nrow(ihd)

# mi
mi <- mi %>%
    filter(!has_mi == 0)
nrow(mi)
mi <- mi %>%
    group_by(person_id) %>%
    slice_min(date_mi_diagnosis, with_ties = FALSE)
nrow(mi)

# pvd
pvd <- pvd %>%
    filter(!has_pvd == 0)
nrow(pvd)
pvd <- pvd %>%
    group_by(person_id) %>%
    slice_min(date_pvd_diagnosis, with_ties = FALSE)
nrow(pvd)

# stroke
stroke <- stroke %>%
    filter(!has_stroke == 0)
nrow(stroke)
stroke <- stroke %>%
    group_by(person_id) %>%
    slice_min(date_stroke_diagnosis, with_ties = FALSE)
nrow(stroke)

# valve_disease
valve_disease <- valve_disease %>%
    filter(!has_valve_disease == 0)
nrow(valve_disease)
valve_disease <- valve_disease %>%
    group_by(person_id) %>%
    slice_min(date_valve_disease_diagnosis, with_ties = FALSE)
nrow(valve_disease)

head(angina)
head(cardiomyopathy)

# join all CVDs
cvd <- full_join(angina, cardiomyopathy, by = "person_id")
cvd <- full_join(cvd, heart_failure, by = "person_id")
cvd <- full_join(cvd, hypertension, by = "person_id")
cvd <- full_join(cvd, ihd, by = "person_id")
cvd <- full_join(cvd, mi, by = "person_id")
cvd <- full_join(cvd, pvd, by = "person_id")
cvd <- full_join(cvd, stroke, by = "person_id")
cvd <- full_join(cvd, valve_disease, by = "person_id")

cvd <- cvd %>%
    mutate_at(c('has_angina','has_cardiomyopathy', 'has_heart_failure', 'has_hypertension', 'has_ihd','has_mi', 'has_pvd', 'has_stroke', 'has_valve_disease'), ~replace_na(.,0))
head(cvd)

nrow(cvd)

sum(duplicated(cvd))

# save dataset
# set destination table - do this regardless of whether it already exists
dest_dataset <- bq_dataset("yhcr-prd-phm-bia-core","CB_MYSPACE_AH")
dest_table <- bq_table(dest_dataset, "All_CVD_First_Occurrence")

# create a full table spec as follows
# then pass this to bq_table_upload
dest_fields <- bq_fields(list(bq_field("person_id", "INT64", "REQUIRED"),
                              bq_field("date_angina_diagnosis", "DATE"),
                              bq_field("has_angina", "INT64"),
                              bq_field("date_cardiomyopathy_diagnosis", "DATE"),
                              bq_field("has_cardiomyopathy", "INT64"),
                              bq_field("date_heart_failure_diagnosis", "DATE"),
                              bq_field("has_heart_failure", "INT64"),
                              bq_field("date_hypertension_diagnosis", "DATE"),
                              bq_field("has_hypertension", "INT64"),
                              bq_field("date_ihd_diagnosis", "DATE"),
                              bq_field("has_ihd", "INT64"),
                              bq_field("date_mi_diagnosis", "DATE"),
                              bq_field("has_mi", "INT64"),
                              bq_field("date_pvd_diagnosis", "DATE"),
                              bq_field("has_pvd", "INT64"),
                              bq_field("date_stroke_diagnosis", "DATE"),
                              bq_field("has_stroke", "INT64"),
                              bq_field("date_valve_disease_diagnosis", "DATE"),
                              bq_field("has_valve_disease", "INT64")))

# delete an existing table if necessary
if(bq_table_exists(dest_table)) bq_table_delete(dest_table)
# upload file_data to dest_table
bq_table_upload(dest_table, cvd, fields = dest_fields, create_disposition='CREATE_IF_NEEDED', write_disposition='WRITE_TRUNCATE')
