==> code/91.Inpatient_LOS_CVD.ipynb <==
library(bigrquery)
library(tidyverse)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# CVD Episodes
sql3 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Cardiovascular_Diseases_Episodes`"
tb3 <- bq_project_query(projectid, sql3)
cvd_ep <-bq_table_download(tb3)

# Diabetes
sql1 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Diabetes`"
tb1 <- bq_project_query(projectid, sql1)
diabetes <-bq_table_download(tb1)

colnames(cvd_ep)

# Keep only relevant variables, choose earliest CVD date and LOS for that CVD
cvd <- cvd_ep %>%
    mutate(earliest_cvd_date = pmin(date_angina_diagnosis, date_cardiomyopathy_diagnosis, date_heart_failure_diagnosis, date_hypertension_diagnosis, date_ihd_diagnosis, date_mi_diagnosis, date_pvd_diagnosis, date_stroke_diagnosis, date_valve_disease_diagnosis, na.rm = TRUE)) %>%
    mutate(age_earliest_cvd = case_when(
        earliest_cvd_date == date_angina_diagnosis ~ age_angina_diag,
        earliest_cvd_date == date_cardiomyopathy_diagnosis ~ age_cardiomyopathy_diag,
        earliest_cvd_date == date_heart_failure_diagnosis ~ age_heart_failure_diag,
        earliest_cvd_date == date_hypertension_diagnosis ~ age_hypertension_diag,
        earliest_cvd_date == date_ihd_diagnosis ~ age_ihd_diag,
        earliest_cvd_date == date_mi_diagnosis ~ age_mi_diag,
        earliest_cvd_date == date_pvd_diagnosis ~ age_pvd_diag,
        earliest_cvd_date == date_stroke_diagnosis ~ age_stroke_diag,
        earliest_cvd_date == date_valve_disease_diagnosis ~ age_valve_disease_diag)) %>%
    mutate(los_earliest_cvd = case_when(
        earliest_cvd_date == date_angina_diagnosis ~ los_angina_days,
        earliest_cvd_date == date_cardiomyopathy_diagnosis ~ los_cardiomyopathy_days,
        earliest_cvd_date == date_heart_failure_diagnosis ~ los_heart_failure_days,
        earliest_cvd_date == date_hypertension_diagnosis ~ los_hypertension_days,
        earliest_cvd_date == date_ihd_diagnosis ~ los_ihd_days,
        earliest_cvd_date == date_mi_diagnosis ~ los_mi_days,
        earliest_cvd_date == date_pvd_diagnosis ~ los_pvd_days,
        earliest_cvd_date == date_stroke_diagnosis ~ los_stroke_days,
        earliest_cvd_date == date_valve_disease_diagnosis ~ los_valve_disease_days)) %>%
    select(person_id, has_cvd, cvd_number, earliest_cvd_date, los_earliest_cvd, age_earliest_cvd)

# Keep relevant vars in diabetes 
diab <- diabetes %>%
    select(person_id, age_earliest_diab, earliest_diabetes_diag, any_diabetes)

head(cvd)
head(diab)

# Join cvd and diab - inner join, left join or full join?
db_cvd <- inner_join(diab, cvd, by = "person_id")
colnames(db_cvd)

sum(db_cvd$los_earliest_cvd == 1)

exclude_1_day <- db_cvd %>%
    filter(!los_earliest_cvd <= 1)

summary(exclude_1_day$los_earliest_cvd)
IQR(exclude_1_day$los_earliest_cvd)
