==> code/92.Inpatient_LOS_Diabetes.ipynb <==
library(bigrquery)
library(tidyverse)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# Diabetes
sql1 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Diabetes`"
tb1 <- bq_project_query(projectid, sql1)
diabetes <-bq_table_download(tb1)
diab <- diabetes %>% select(person_id, any_diabetes)

# Secondary Care Episodes table
sql2 <- "SELECT person_id, episode_start_date, episode_end_date FROM `yhcr-prd-phm-bia-core.CB_FDM_Warehouse_V2.tbl_episode` 
WHERE episode_start_date BETWEEN '2002-01-01' AND '2022-12-31'"
tb2 <- bq_project_query(projectid, sql2)
episodes <-bq_table_download(tb2)

# Join diabetes and episodes tables
diab_epi <- left_join(diab, episodes, by = "person_id")
diab_epi <- diab_epi %>% filter(!is.na(episode_start_date))
head(diab_epi)

# Calculate length of stay
diab_epi <- diab_epi %>%
    mutate(los_all_days = round(difftime(episode_end_date, episode_start_date, units = "days"))) %>%
    mutate(los_all_days = as.integer(los_all_days))

# Keep only one row per patient? Based on?

table(diab_epi$los_all_days == 1)

dep <- diab_epi %>% 
    filter(diab_epi$los_all_days >= 2)

summary(dep$los_all_days)
IQR(dep$los_all_days)
