==> code/94.Diabetes_Death_Certificate.ipynb <==
library(bigrquery)
library(tidyverse)

# Store the project ID
projectid = "yhcr-prd-phm-bia-core"

# Diabetes
sql1 <- "SELECT *  FROM `yhcr-prd-phm-bia-core.CB_MYSPACE_AH.Diabetes`"
tb1 <- bq_project_query(projectid, sql1)
diabetes <-bq_table_download(tb1)

# Import and Search Death certificate dataset for ICD10 CVD death codes
# Type 1 Diabetes
sql8 <- "SELECT * 
FROM `yhcr-prd-phm-bia-core.CB_FDM_DeathCertificates.tbl_NEC_Deaths_QWO_20230817_RELEASE`
WHERE s_cod_code_1 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_2 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_3 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_4 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_5 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_6 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_7 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_8 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_9 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_10 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_11 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_12 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_13 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_14 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')
OR s_cod_code_15 IN ('E101','O240','E872','E100','O240','R402','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240','E104D','O240','E107','O240','E105','O240','L984','E105','O240','R02X','E103D','O240D','H360A','H360A','E103D','O240D','E109','E10','O240','E109','E10','O240','O240','E109','E10','O240','E106','O240','E109','E10','O240','E109','O240','E104D','O240D','G590A','G590A','E104D','O240D','E104D','O240D','G632A','G632A','E104D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E100','O240','R402','E103D','O240D','H280A','H280A','E103D','O240D','E105D','O240D','I792A','I792A','E105D','O240D','E106D','O240D','M142A','M142A','E106D','O240D','E106D','O240D','M146A','M146A','E106D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E102D','O240D','N083A','N083A','E102D','O240D','E103D','O240D','H360A','H360A','E103D','O240D','E106D','O240D','K318A')"
tb8 <- bq_project_query(projectid, sql8)
diab1_dcert <- bq_table_download(tb8)

# Type 2 Diabetes
sql81 <- "SELECT * 
FROM `yhcr-prd-phm-bia-core.CB_FDM_DeathCertificates.tbl_NEC_Deaths_QWO_20230817_RELEASE`
WHERE s_cod_code_1 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_2 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_3 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_4 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_5 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_6 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_7 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_8 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_9 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_10 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_11 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_12 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_13 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_14 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')
OR s_cod_code_15 IN ('E111','O241','E872','E110','O241','R402','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241','E114D','O241','E117','O241','E115','O241','L984','E115','O241','R02X','E113D','O241D','H360A','H360A','E113D','O241D','E119','E11','O241','O241','E119','E11','O241','E119','E11','O241','E119','O241','E114D','G590A','G590A','E114D','E114D','O241D','G632A','G632A','E114D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E110','O241','R402','E113D','O241D','H280A','H280A','E113D','O241D','E115D','O241D','I792A','I792A','E115D','O241D','E116D','O241D','M142A','M142A','E116D','O241D','E116D','O241D','M146A','M146A','E116D','O241D','NA','NA','E116','O241','E870','E112D','O241D','N083A','N083A','E112D','O241D','E112D','O241D','N083A','N083A','E112D','O241D','E113D','O241D','H360A','H360A','E113D','O241D','E116','O241','K318')"
tb81 <- bq_project_query(projectid, sql81)
diab2_dcert <- bq_table_download(tb81)

# Add has_*_dc variables
diab1_dcert <- diab1_dcert %>%
    select(person_id) %>%
    mutate(has_t1d_dc = 1)

diab2_dcert <- diab2_dcert %>%
    select(person_id) %>%
    mutate(has_t2d_dc = 1)

# Keep only person_id and any_diabetes
diab <- diabetes %>%
    select(person_id, any_diabetes)

ddc <- left_join(diab, diab1_dcert, by = "person_id")
ddc <- left_join(ddc, diab2_dcert, by = "person_id")
ddc <- ddc %>%
    mutate(has_t1d_dc = ifelse(is.na(has_t1d_dc),0,1),
          has_t2d_dc = ifelse(is.na(has_t2d_dc),0,1))
table(ddc$has_t1d_dc)
table(ddc$has_t2d_dc)
table(ddc$has_t1d_dc, ddc$has_t2d_dc)

# Number (%) of deceased diabetic patients who have diabetes on their death certificate
# T1D
(62/10528)*100

# T2D
(2415/10528)*100

# Both
(9/10528)*100
